% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_count.R
\name{add_count}
\alias{add_count}
\title{Adds the count to a specific variable}
\usage{
add_count(d, variable_name)
}
\arguments{
\item{d}{a tibble or dataframe}

\item{variable_name}{the column with counts wanted}
}
\value{
a tibble with the counts appended to the `variable_name` column
}
\description{
adds the count of each variable to its value
}
\examples{
\dontrun{
# Adds the count of each animal to its animal listing in the column animal
library(tidyverse)
d <- tibble(animal = sample(c("cat", "dog", "mouse", "rabbit"), 20, TRUE),
  other_var = rnorm(20))

add_count(d, animal)
}
}
