% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connections.R
\name{add_SAM_utility_cols}
\alias{add_SAM_utility_cols}
\title{Add SAM utility columns to table}
\usage{
add_SAM_utility_cols(d)
}
\arguments{
\item{d}{A dataframe}
}
\value{
A dataframe with three additional columns
}
\description{
When working in a Health Catalyst Source Area Mart (SAM),
utility columns are added automatically when running a non-R binding
}
\examples{
d <- data.frame(a = c(1,2,NA,NA),
                 b = c(100,300,200,150))
d <- add_SAM_utility_cols(d)

}
