% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_list_constructors.R
\name{as.model_list}
\alias{as.model_list}
\title{Make models into model_list object}
\usage{
as.model_list(..., listed_models = NULL, target = ".outcome", model_class)
}
\arguments{
\item{...}{\code{caret}-trained models to put into a model list}

\item{listed_models}{Use this if your models are already in a list}

\item{target}{Quoted name of response variable}

\item{model_class}{"classification" or "regression"}
}
\value{
A model_list with child class type_list
}
\description{
Make models into model_list object
}
