% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-kmeans.R
\name{assignClusterLabels}
\alias{assignClusterLabels}
\title{Assign labels to the clusters}
\usage{
assignClusterLabels(cm, k)
}
\arguments{
\item{cm}{A confusion matrix, with each row converting to percentage}

\item{k}{A number, number of clusters}
}
\value{
A list of length k, showing the labels that are assigned to each cluster.
}
\description{
compute the distance of a point from a line given the 
x, y axes of the point and the slope and intercept of the line.
}
\examples{
data(iris)
head(iris)
kmeans.fit <- kmeans(iris[,1:4],3) 
labs <- iris[,5]
cls <- kmeans.fit[["cluster"]]
cm <- calculateConfusion(labels = labs, clusters = cls)
assignClusterLabels(cm, 3)

}
\references{
\url{http://healthcare.ai}
}
\seealso{
\code{\link{healthcareai}}
}
