% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_db_connections_build.R
\name{build_connection_string}
\alias{build_connection_string}
\title{Build a connection string for use with MSSQL and dbConnect}
\usage{
build_connection_string(server, driver = "SQL Server", database = NA,
  trusted = TRUE, user_id = NA, password = NA)
}
\arguments{
\item{server}{Character, required. The name of the server you are trying to
connect to. Quoted.}

\item{driver}{Character, optional. Defaults to "SQL Server"}

\item{database}{Character, optional. If provided, connection string will
include a specific database. If NA (default), it will connect to master and
you'll have to specify the database when running a query.}

\item{trusted}{Logical, optional, defaults to TRUE.}

\item{user_id}{Character, optional. Leave blank if using trusted.}

\item{password}{Character, optional. Leave blank if using trusted.}
}
\value{
A connection string
}
\description{
Handy utility to build a connection string to pass into
\code{DBI::dbConnect}. Accepts trusted connections or username/password.
}
\examples{
my_con <- build_connection_string(server = "localhost")
# con <- DBI::dbConnect(odbc::odbc(), .connection_string = my_con)

# with username and password
my_con <- build_connection_string(server = "localhost",
                                  user_id = "jules.winnfield",
                                  password = "pathoftherighteous")
# con <- DBI::dbConnect(odbc::odbc(), .connection_string = my_con)

}
