% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-process-variable-recommendations.R
\name{build_one_level_df}
\alias{build_one_level_df}
\title{Replace all value in the column of a dataframe with a given value.}
\usage{
build_one_level_df(dataframe, modifiable_variable, level)
}
\arguments{
\item{dataframe}{a dataframe}

\item{modifiable_variable}{The name of the column whose values we wish to
change}

\item{level}{The value to use in the \code{modifiable_variable} column}
}
\value{
A dataframe
}
\description{
Takes a dataframe and replaces all the values in the 
\code{modifiable_variable} column with the value \code{level}. Columns
tracking which variable and level were used are also added to the dataframe.
}
\keyword{internal}
