% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-process-variable-recommendations.R
\name{build_process_variable_df_list}
\alias{build_process_variable_df_list}
\title{Build a list of dataframes with new predictions for each modifiable variable}
\usage{
build_process_variable_df_list(dataframe, grain_column_values,
  modifiable_variable_levels, predict_function, low_probabilities_desired)
}
\arguments{
\item{dataframe}{a dataframe consisting of deployment data}

\item{grain_column_values}{the grain column (values, not just the name) of 
the deployment data}

\item{modifiable_variable_levels}{a list whose}

\item{predict_function}{a function with which to make new predictions on the
data}

\item{low_probabilities_desired}{a boolean indicating whether the goal is to
increase or decrease the predictions/predicted probabilities}
}
\value{
a list of dataframes
}
\description{
Builds a list consisting of one dataframe per row in
\code{dataframe} which shows how the predictions for that row change as
each of the mofifiable variables are altered
}
\keyword{internal}
