% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-LIME.R
\name{buildLIMEExplainer}
\alias{buildLIMEExplainer}
\title{Build a LIME explainer for a trained classification model}
\usage{
buildLIMEExplainer(trainingSet, trainedModel, binContinuous = TRUE,
  bins = 5)
}
\arguments{
\item{trainingSet}{The data set on which the model was trained}

\item{trainedModel}{The trained model}

\item{bins}{The number of bins to use for continuous variables if 
binContinuous is TRUE}
}
\description{
Build a LIME explainer for a trained classification model 
using the trained model and the training set used to train the model
}
