% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{CalculateAllCorrelations}
\alias{CalculateAllCorrelations}
\title{Correlation analysis on an input table over all numeric columns}
\usage{
CalculateAllCorrelations(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
df A dataframe with column names and corresponding correlations
with the target column
}
\description{
Calculate correlations between every numeric column in a table
}
\examples{
library(HCRTools)

df <- data.frame(a=c(1,2,3,4,5,6),
b=c(6,5,4,3,2,1),
c=c(3,4,2,1,3,5),
d=c('M','F','F','F','M','F')) #<- is ignored

res <- CalculateAllCorrelations(df)
res
}
\references{
\url{https://community.healthcatalyst.com/community/data-science}
}
\seealso{
\code{\link{HCRTools}}
}

