% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-variation.R
\name{calculateCOV}
\alias{calculateCOV}
\title{Calculate coefficient of variation}
\usage{
calculateCOV(vector)
}
\arguments{
\item{vector}{A vector of numbers}
}
\value{
A scalar
}
\description{
Find coefficient of variation by dividing vector standard 
deviation by the mean and muliplying by 100.
}
\examples{
df <- data.frame(a = c(1,2,NA,NA),
                 b = c(100,300,200,150))
res1 <- calculateCOV(df$a)
res1

res2 <- calculateCOV(df$b)
res2
}
\references{
\url{http://healthcare.ai}
}
\seealso{
\code{\link{healthcareai}}
}
