% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{calculateHalfSD}
\alias{calculateHalfSD}
\title{Calculate half std deviation up/down for each numeric field in row}
\usage{
calculateHalfSD(r, returnOriginal = FALSE)
}
\arguments{
\item{r}{Single row coming from a data frame}
}
\description{
Add/subtract each numeric col (for each row) by half std dev
such that we have a new alternate data frame
}
\references{
\url{http://hctools.org/}
}
\seealso{
\code{\link{HCRTools}}
}

