% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-feature-availability-profiler.R
\name{calculateHourBins}
\alias{calculateHourBins}
\title{Calculate a vector of reasonable time bins}
\usage{
calculateHourBins(lastHourOfInterest)
}
\arguments{
\item{lastHourOfInterest}{Number (hour) scalar representing the last hour of 
interest}
}
\value{
numeric vector of hours, reasonably spaced
}
\description{
Given a number of hours, generate a reasonable vector of bins 
in hours such that the first day is divided into multiple days are divided 
into 24 h bins up to 90 days worth. Typically used with 
featureAvailabilityProfiler
}
\references{
\url{http://healthcareai-r.readthedocs.io}
}
\seealso{
\code{\link{healthcareai}}
result <- calculateHourBins(90)
result
}
