% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-model-performance.R
\name{calculatePerformance}
\alias{calculatePerformance}
\title{Generate performance metrics after model has been trained}
\usage{
calculatePerformance(predictions, ytest, type)
}
\arguments{
\item{predictions}{A vector of predictions from a machine learning model.}

\item{ytest}{A vector of the true labels. Must be the same length as 
predictions.}

\item{type}{A string. Indicates model type and can be "regression" or 
"classification". 
Defaults to SS.}
}
\value{
Curves (if classification); otherwise nothing. Prints results.
}
\description{
Generates AU_ROC and AU_PR (including 95% confidence)
}
\references{
\url{http://healthcareai-r.readthedocs.io}
}
\seealso{
\code{\link{healthcareai}}
}
