% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{CalculateTargetedCorrelations}
\alias{CalculateTargetedCorrelations}
\title{Correlation analysis on an input table, focusing on one target variable}
\usage{
CalculateTargetedCorrelations(df, targetcol)
}
\arguments{
\item{df}{A dataframe}

\item{target.col}{Name of target column against which correlations will be
calculated}
}
\value{
df A dataframe with column names and corresponding correlations and
p-values with the target column
}
\description{
Calculates correlations between each numeric column in a table
and and a target column
}
\examples{
library(HCRTools)

df <- data.frame(a=c(1,2,3,4,5,6),
b=c(6,5,4,3,2,1),
c=c(3,4,2,1,3,5),
d=c('M','F','F','F','M','F')) <- is ignored

res <- CalculateTargetedCorrelations(df,'c')
res
}
\references{
\url{https://community.healthcatalyst.com/community/data-science}
}
\seealso{
\code{\link{HCRTools}}
}

