% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{calulcateAlternatePredictions}
\alias{calulcateAlternatePredictions}
\title{Recalculate predicted value based on alternate scenarios}
\usage{
calulcateAlternatePredictions(df, modelObj, type, outVectorAppend = NULL,
  removeCols = NULL)
}
\arguments{
\item{df}{Data frame from which we'll calculate alternate predictions}

\item{modelObj}{Object representing the model that is used for predictions}

\item{type}{String representing which type of model is used}

\item{outVectorAppend}{Optional list of values that we'll append predictions
to. If not used, then a new vector is created.}
}
\description{
After getting alternate features via calculateSDChanges
recalculate predicted values for each row in df.
}
\examples{
library(caret)
X <- data.frame(a=c(1,2,3,1),
                b=c('m','f','f','m'),
                c=c(0.7,1.4,2.4,2.0),
                d=c(100,250,200,150))

y <- c('y','n','y','n')

X_alt <- calculateSDChanges(df = X,
                            rowToBeAltered = X[4,],
                            sizeOfSDPerturb = 0.5,
                            colsToAlter = c('a','c'))

X_alt

combined = cbind(X,y)

glmOb <- glm(y ~ a + b + c + d,
             data = combined,
             family = binomial(link = "logit"))

glmOb

originalPred <- predict(object = glmOb,
                        newdata = X[4,],
                        type = 'response')
originalPred

alternatePred <- calulcateAlternatePredictions(df = X_alt,
                                               modelObj = glmOb,
                                               type = 'lasso',
                                               removeCols = 'AlteredCol')
alternatePred
}
\references{
\url{http://healthcare.ai}
}
\seealso{
\code{\link{healthcareai}}
}

