% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-logHelpers.R
\name{catalyst_test_deploy_in_prod}
\alias{catalyst_test_deploy_in_prod}
\title{Test function to check that the production environment is active.}
\usage{
catalyst_test_deploy_in_prod()
}
\description{
This is meant to test the production environment on ETL servers. 
Use it instead of a long running deploy script.
}
\seealso{
\code{\link{start_prod_logs}}

\code{\link{stop_prod_logs}}
}
