% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data_utils.R
\name{convert_date_cols}
\alias{convert_date_cols}
\title{Prepare data for machine learning}
\usage{
convert_date_cols(d)
}
\arguments{
\item{d}{A dataframe or tibble containing data to try to convert to dates.}
}
\value{
Prepared data frame with reusable recipe object for future data
  preparation in attribute "recipe". Attribute recipe contains the names of
  ignored columns (those passed to ...) in attribute "ignored_columns".
}
\description{
in the `recipe`
  attribute of the output data frame. If a recipe object is passed to
  `prep_data` via the `recipe` argument, that recipe will be applied to the
  data. This allows you to transform data in model training and apply exactly
  the same transformations in model testing and deployment. The new data must
  be identical in structure to the data that the recipe was prepared with.
}
\examples{
d_train <- pima_diabetes[1:700, ]
d_test <- pima_diabetes[701:768, ]

}
\seealso{
\code{\link{hcai_impute}}, \code{\link{tune_models}},
  \code{\link{predict.model_list}}
}
