% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingness.R
\name{countMissingData}
\alias{countMissingData}
\title{Function to find proportion of NAs in each column of a dataframe or matrix}
\usage{
countMissingData(x, userNAs = NULL)
}
\arguments{
\item{x}{A data frame or matrix}

\item{userNAs}{A vector of user defined NA values.}
}
\description{
DEPRICATED. Use \code{\link{missingness}} instead.
}
