% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{CountPercentEmpty}
\alias{CountPercentEmpty}
\title{Calculates percentage of each column in df that is NULL (NA)}
\usage{
CountPercentEmpty(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
colList A vector that contains the names of the columns with greater
than 50 categories
}
\description{
Returns a vector with percentage of each column that is NULL
in the original data frame
}
\examples{
df = data.frame(a=c(1,2,NA,NA,3),
                b=c(NA,NA,NA,NA,NA),
                c=c(NA,NA,'F','M',NA))
df.result = CountPercentEmpty(df)
df.result
}
\references{
\url{https://community.healthcatalyst.com/community/data-science}
}
\seealso{
\code{\link{HCRTools}}
}

