% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-variation.R
\name{createCombinations}
\alias{createCombinations}
\title{Find all possible combinations}
\usage{
createCombinations(categoricalCols)
}
\arguments{
\item{categoricalCols}{A vector of strings}
}
\value{
A list of sub-lists. Each sub-list represents one possible 
combination.
}
\description{
For a given vector of strings, find all possible combinations 
of those strings.
}
\examples{
categoricalCols <- c("LactateOrderHospital",
                     "LactateOrderProvSpecialtyDSC",
                     "LactateOrderProvNM")
y <- createCombinations(categoricalCols)

# Let's look at one possible combination
print(unlist(y[3]))
}
\references{
\url{http://healthcare.ai}
}
\seealso{
\code{\link{healthcareai}} \code{\link{findVariation}} 
\code{\link{createVarianceTallTable}}
}
