% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-variation.R
\name{createVarianceTallTable}
\alias{createVarianceTallTable}
\title{Transform a dataframe to be three columns and tall instead of wide}
\usage{
createVarianceTallTable(df, categoricalCols, measureCol)
}
\arguments{
\item{df}{A dataframe}

\item{categoricalCols}{Vector of categorical columns}

\item{measureCol}{A string of the measure column of interest}
}
\value{
A dataframe of three columns
}
\description{
When dealing with a table that could be unexpectedly wide,
it helps to instead fix it at three columns and let it get tall. Since this 
provides fixed columns, it's easy to put in a database table.
}
\examples{
df <- data.frame(LactateOrderProvSpecialtyDSC = c("Pulmonary Disease",
                                                  "Family Medicine"),
                 LactateOrderProvNM = c("Hector Salamanca",
                                        "Gus Fring"),
                 LactateResultNBR = c(43.4,35.1))

head(df)

categoricalCols <- c("LactateOrderProvSpecialtyDSC","LactateOrderProvNM")
measureCol <- "LactateResultNBR"
dfResult <- createVarianceTallTable(df, categoricalCols, measureCol)
head(dfResult)
}
\references{
\url{http://healthcare.ai}
}
\seealso{
\code{\link{healthcareai}} 
\code{\link{findVariation}}
}
