% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-kmeans.R
\name{dataScale}
\alias{dataScale}
\title{Center and scale columns in a numeric data frame}
\usage{
dataScale(df)
}
\arguments{
\item{df}{A numeric data frame}
}
\value{
A list that contains a vector of column means, a vector of column
sds, and a scaled data frame.
}
\description{
center and scale columns in a numeric data frame using means and 
standard deviations
}
\examples{
df <- data.frame(a = c(2,1,3,2,4),b = c(NA,8,6,7,9))
res <- dataScale(df)
res
}
\references{
\url{http://healthcare.ai}
}
\seealso{
\code{\link{healthcareai}}
}
