% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_db_connections_write.R
\name{db_write}
\alias{db_write}
\title{Write to an MSSQL database table}
\usage{
db_write(d, con, table_name, schema = "dbo", overwrite = FALSE)
}
\arguments{
\item{d}{A dataframe or tibble to write to the database. Required.}

\item{con}{An RODBC database connection. Can be made using
\code{RODBC::odbcDriverConnect}. Required.}

\item{table_name}{Character, required.}

\item{schema}{Character, optional. Defaults to "dbo".}

\item{overwrite}{Logical, optional, defaults to FALSE.}
}
\value{
Silently returns number of rows written to new table.
}
\description{
Use a database connection to write to a table in a specified
schema.
}
\details{
This function currently requires an RODBC connection to a specific
database. See examples for how to generate one. Column names and types in d
must match those in destination table.
}
\examples{
\dontrun{
my_con <- build_connection_string(server = "localhost",
                                  database = "myDB")
rcon <- RODBC::odbcDriverConnect(my_con)

db_write(conn)
}
}
