% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeployLinearMixedModel.R
\docType{class}
\name{DeployLinearMixedModel}
\alias{DeployLinearMixedModel}
\title{Deploy a production-ready predictive RandomForest model}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DeployLinearMixedModel
}
\arguments{
\item{type}{The type of model (either 'regression' or 'classification')}

\item{df}{Dataframe whose columns are used for calc.}

\item{grainCol}{The data frame's column that has IDs pertaining to the grain}

\item{personCol}{The data frame's columns that represents the patient/person}

\item{testWindowCol}{This column dictates the split between model training and
test sets. Those rows with zeros in this column indicate the training set
while those that have ones indicate the test set}

\item{predictedCol}{Column that you want to predict.}

\item{impute}{For training df, set all-column imputation to F or T.
This uses mean replacement for numeric columns
and most frequent for factorized columns.
F leads to removal of rows containing NULLs.}

\item{debug}{Provides the user extended output to the console, in order
to monitor the calculations throughout. Use T or F.}
}
\description{
This step allows one to
\itemize{
\item Create a final model on all of your training data
\item Automatically save the model
\item Run the model against test data to generate predictions
\item Push these predictions to SQL Server
}
}
\seealso{
\code{\link{HCRTools}}
}
\keyword{datasets}

