% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_supervised_model.R
\docType{class}
\name{DeploySupervisedModel}
\alias{DeploySupervisedModel}
\title{Deploy a production-ready predictive model}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DeploySupervisedModel
}
\arguments{
\item{type}{The type of model (either 'regression' or 'classification')}

\item{df}{Dataframe whose columns are used for calc.}

\item{grain.col}{The dataframe's column that has IDs pertaining to the grain}

\item{window.col}{This column dictates the split between model training and
test sets. Those rows with zeros in this column indicate the training set
while those that have ones indicate the test set}

\item{predicted.col}{Column that you want to predict.}

\item{impute}{For training df, set all-column imputation to F or T.
This uses mean replacement for numeric columns
and most frequent for factorized columns.
F leads to removal of rows containing NULLs.}

\item{debug}{Provides the user extended output to the console, in order
to monitor the calculations throughout. Use T or F.}
}
\description{
This step allows one to
\itemize{
\item Create a final model on all of your training data
\item Automatically save the model
\item Run the model against test data to generate predictions
\item Push these predictions to SQL Server
}
}
\examples{
# The data will read in as-is and you can find the data itself here
# C:\\Users\\levi.thatcher\\Documents\\R\\win-library\\3.2\\HCRTools\\extdata OR
# C:\\Program Files\\R\\R-3.2.3\\library\\HCRTools\\extdata

#### Classification example using data from csv ####
# This example requires
#    1) You set your working directory to source file location
#    2) To receive predictions from R back to SQL Server, you'll need to save
#       and run an entity in SAMD that has the following columns
#
# GrainID decimal(38,0) not null, <--change col to match ID in summary table
# PredictedProbNBR decimal(38,2),
# Factor1TXT varchar(255),
# Factor2TXT varchar(255),
# Factor3TXT varchar(255),

# If you prefer to not use SAMD, execute this in SSMS to create output table:
# CREATE TABLE dbo.HCRDeployTest1BASE(
#   BindingID float, BindingNM varchar(255), LastLoadDTS datetime2,
#   GrainID int <--change to match inputID, PredictedProbNBR decimal(38, 2),
#   Factor1TXT varchar(255), Factor2TXT varchar(255), Factor3TXT varchar(255)
# )

#setwd("C:/Your/script/location") # Needed if using YOUR CSV file
ptm <- proc.time()
library(HCRTools)

connection.string <- 'driver={SQL Server};
                      server=localhost;
                      database=SAM;
                      trusted_connection=true'

# Can delete this line in your work:
csvfile <- system.file("extdata", "HREmployeeDeploy.csv",package = "HCRTools")

totaldf <- read.csv(file = csvfile, #<--replace with 'your/path'
                    header = TRUE,
                    na.strings = 'NULL')

totaldf$VacationHours <- NULL

head(totaldf)

o <- DeploySupervisedModel$new(type = 'classification',
                               df = totaldf,
                               grain.col = 'GrainID',
                               window.col = 'InWindow',
                               predicted.col = 'SalariedFlag',
                               impute = TRUE,
                               debug = FALSE,  # <-- change to TRUE to debug
                               use.saved.model = FALSE)


o$deploy(model = 'rf',
         cores = 1,
         sqlcnxn = connection.string,
         # Note: Do not use [ or ] in output table
         dest.schema.table = 'dbo.HCRDeployTest1BASE',
         debug = FALSE,  # <-- change this to TRUE to debug
)

print(proc.time() - ptm)


#### Regression example using data from SQL Server ####
# This example requires
#     1) You set your working directory to source file location
#     2) To receive predictions from R back to SQL Server, you'll need to
#        save and run an entity in SAMD that has the following columns

# GrainID decimal(38,0) not null, <--change col to match ID in summary table
# PredictedValueNBR decimal(38,2),
# Factor1TXT varchar(255),
# Factor2TXT varchar(255),
# Factor3TXT varchar(255),

# If you prefer to not use SAMD, execute this in SSMS to create output table:
# CREATE TABLE dbo.HCRDeployTest2BASE(
#   BindingID float, BindingNM varchar(255), LastLoadDTS datetime2,
#   GrainID int <--change to match inputID, PredictedValueNBR decimal(38, 2),
#   Factor1TXT varchar(255), Factor2TXT varchar(255), Factor3TXT varchar(255)
# )

ptm <- proc.time()
library(HCRTools)

connection.string <- 'driver={SQL Server};
                      server=localhost;
                      database=SAM;
                      trusted_connection=true'

# Use this for an example SQL source:
# query <- "SELECT * FROM [SAM].[YourCoolSAM].[SomeTrainingSetTable]"
# totaldf <- SelectData(connection.string, query)

# Can delete these four lines when you set up your SQL connection/query
csvfile <- system.file("extdata", "HREmployeeDeploy.csv",package = "HCRTools")
totaldf <- read.csv(file = csvfile,
                    header = TRUE,
                    na.strings = 'NULL')

head(totaldf)

totaldf <- subset(totaldf, select = -c(SalariedFlag))

o <- DeploySupervisedModel$new(type = 'regression',
                               df = totaldf,
                               grain.col = 'GrainID',
                               window.col = 'InWindow',
                               predicted.col = 'VacationHours',
                               impute = TRUE,
                               debug = FALSE,  # <-- change to T for debug
                               use.saved.model = FALSE)

o$deploy(model = 'logit',
         cores = 1,
         sqlcnxn = connection.string,
         debug = FALSE,  # <-- change this to TRUE to debug
         # Note: Do not use [ or ] in output table
         dest.schema.table = 'dbo.HCRDeployTest2BASE')

print(proc.time() - ptm)

}
\seealso{
\code{\link{DevelopSupervisedModel}}

\code{\link{HCRTools}}
}
\keyword{datasets}

