% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeploySupervisedModelParameters.R
\docType{class}
\name{DeploySupervisedModelParameters}
\alias{DeploySupervisedModelParameters}
\title{DeploySupervisedModelParameters class to set up parameters required to build
DeploySupervisedModel classes}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DeploySupervisedModelParameters
}
\description{
This step allows one to create deploy models on your data
and helps determine which performs best.
}
\references{
\url{http://products.healthcatalyst.com/Predictive}
}
\seealso{
\code{\link{HCRTools}}
}
\keyword{datasets}

