% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/develop_supervised_model.R
\docType{class}
\name{DevelopSupervisedModel}
\alias{DevelopSupervisedModel}
\title{Compare predictive models, created on your data}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DevelopSupervisedModel
}
\arguments{
\item{type}{The type of model (either 'regression' or 'classification')}

\item{df}{Dataframe whose columns are used for calc.}

\item{grain.col}{The dataframe's column that has IDs pertaining to the grain}

\item{predicted.col}{Column that you want to predict.}

\item{impute}{Set all-column imputation to F or T.
This uses mean replacement for numeric columns
and most frequent for factorized columns.
F leads to removal of rows containing NULLs.}

\item{debug}{Provides the user extended output to the console, in order
to monitor the calculations throughout. Use T or F.}
}
\description{
This step allows one to create test models on your data
and helps determine which performs best.
}
\examples{
# The examples will run as-is, but you can find the data used here
# C:\\Users\\levi.thatcher\\Documents\\R\\win-library\\3.2\\HCRTools\\extdata OR
# C:\\Program Files\\R\\R-3.2.3\\library\\HCRTools\\extdata

#### Example using iris dataset ####
ptm <- proc.time()
library(HCRTools)
data(iris)

head(iris)

set.seed(43)
o <- DevelopSupervisedModel$new(type = 'regression',
                                df = iris,
                                #grain.col = 'GrainID',
                                predicted.col = 'Sepal.Width',
                                impute = TRUE)

o$grlasso(var.imp = TRUE,
          cores = 1,
          debug = TRUE)

o$randForest(cores = 1,
             debug = FALSE)
print(proc.time() - ptm)


#### Example using csv data ####
library(HCRTools)
#setwd("C:/Your/script/location") # Needed if using YOUR CSV file
ptm <- proc.time()

# Can delete this line in your work
csvfile <- system.file("extdata", "HREmployeeDev.csv", package = "HCRTools")

totaldf <- read.csv(file = csvfile, #<-- Replace with 'your/path'
                    header = TRUE,
                    na.strings = 'NULL')

head(totaldf)

set.seed(43)
o <- DevelopSupervisedModel$new(type = 'classification',
                                df = totaldf,
                                #grain.col = OPTIONAL/ENCOURAGED
                                predicted.col = 'SalariedFlag',
                                impute = FALSE)

o$grlasso(var.imp = TRUE,
          cores = 1,
          debug = TRUE)


o$randForest(cores = 1,
             debug = FALSE)

o$plotROC()

# For a given true-positive rate, get false-pos rate and 0/1 cutoff
o$getCutOffs('rf', .6)
print(proc.time() - ptm)


#### Example using SQL Server data ####
# This example requires:
#    1) That your local SQL Server has AdventureWorks2012 installed

ptm <- proc.time()
library(HCRTools)
library(RODBC)

connection.string = '
  driver={SQL Server};
  server=localhost;
  database=AdventureWorks2012;
  trusted_connection=true
'
query = "
  SELECT
    [OrganizationLevel]
    ,[MaritalStatus]
    ,[Gender]
    ,[SalariedFlag]
    ,[VacationHours]
    ,[SickLeaveHours]
  FROM [AdventureWorks2012].[HumanResources].[Employee]
"

df <- SelectData(connection.string, query)
head(df)

set.seed(43)
o <- DevelopSupervisedModel$new(type = 'classification',
                                df = df,
                                #grain.col = OPTIONAL/ENCOURAGED,
                                predicted.col = 'MaritalStatus',
                                impute = TRUE)

o$grlasso(cores = 1,
          debug = FALSE)

o$randForest(cores = 1,
             debug = FALSE)

o$plotROC()
print(proc.time() - ptm)

}
\references{
\url{http://products.healthcatalyst.com/Predictive}
}
\seealso{
\code{\link{DeploySupervisedModel}}

\code{\link{HCRTools}}
}
\keyword{datasets}

