% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-kmeans.R
\name{distancePointSegment}
\alias{distancePointSegment}
\title{Compute the distance of a point from a line segment}
\usage{
distancePointSegment(px, py, x1, y1, x2, y2)
}
\arguments{
\item{px}{A numeric vector, x axis of the points}

\item{py}{A numeric vector, y axis of the points}

\item{x1}{A number, x axis of the end point of the line segment}

\item{y1}{A number, y axis of the end point of the line segment}

\item{x2}{A number, x axis of the other end point of the line segment}

\item{y2}{A number, y axis of the other end point of the line segment}
}
\value{
numeric vector, the length between the pointS and the line segment. If the 
intersection point is outside the line segment, return the distance to the nearest 
endpoint.
}
\description{
compute the distance of a point from a line segment given the 
x, y axes of the points.
}
\examples{
distancePointSegment(c(2,0),c(3,2),1,2,3,1)
}
\references{
\url{http://healthcare.ai}

The helper funtion is originally found here
\url{http://paulbourke.net/geometry/pointlineplane/pointline.r}

\url{https://github.com/bryanhanson/ChemoSpecMarkeR/blob/master/R/findElbow.R}
}
\seealso{
\code{\link{healthcareai}}
}
