% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-process-variable-recommendations.R
\name{drop_repeated}
\alias{drop_repeated}
\title{Simultaneously remove duplicate row values in a list of dataframes.}
\usage{
drop_repeated(df_list, column_name)
}
\arguments{
\item{df_list}{A list of dataframes. Each dataframe should contain a column
whose name matches \code{column_name}}

\item{column_name}{The name of the column from which to remove duplicates.}
}
\value{
a list of dataframes with no repeated values in \code{column_name}
}
\description{
Takes a list of dataframes, each of which contains a column 
with the same name and, for each dataframe, removes rows where the value in 
the common column has already occured in a previous row of that dataframe.
}
\keyword{internal}
