% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate.hcai_predicted_df}
\alias{evaluate.hcai_predicted_df}
\title{Get performance metrics from prediction data frame}
\usage{
evaluate.hcai_predicted_df(df)
}
\arguments{
\item{df}{}
}
\description{
Get performance metrics from prediction data frame
}
