% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate_classification}
\alias{evaluate_classification}
\title{Get performance metrics for classification models}
\usage{
evaluate_classification(predicted, actual)
}
\arguments{
\item{predicted}{Vector of predicted probabilities}

\item{actual}{Vector of realized outcomes, must be 0/1}
}
\value{
Numeric vector of scores with metric as names
}
\description{
Get performance metrics for classification models
}
