% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate_regression}
\alias{evaluate_regression}
\title{Get performance metrics for regression predictions}
\usage{
evaluate_regression(predicted, actual)
}
\arguments{
\item{predicted}{Vector of predicted values}

\item{actual}{Vector of realized values}
}
\value{
Numeric vector of scores with metric as names
}
\description{
Get performance metrics for regression predictions
}
\examples{
evaluate_regression(c(2, 4, 6), c(1.5, 4.1, 6.2))
}
