% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_interactions.R
\name{find_interactions}
\alias{find_interactions}
\title{Calculates the h-statistic for either all combinations or for the most
significant interactions.}
\usage{
find_interactions(d, outcome, brute_force = TRUE, order = 3, random_seed,
  verbose = TRUE, greedy_cutoff = 0.05)
}
\arguments{
\item{d}{Dataframe or tibble}

\item{outcome}{Unquoted column name that indicates the target variable.}

\item{brute_force}{Logical. When TRUE (default) every combination of
variables will be searched. If FALSE, a greedy implementation will be used.
If greedy is used hidden interactions will not be found.}

\item{order}{Integer (default = 3) specifying how many orders to search for
in the data set. This function is designed for 2 or 3 orders.}

\item{random_seed}{Numeric. Sets the random seed for the gbm model. If not
provided, models will be randomized.}

\item{verbose}{Logical. If TRUE (default) will list the amount of time it
takes to run brute_force.}

\item{greedy_cutoff}{Numeric. If brute_force is FALSE, all H-statistics above
this threshold will not be explored for further interactions.}
}
\description{
Calculates the h-statistic for either all combinations or for the most
significant interactions.
}
\details{
get_interactions is dependent an tidyverse, gbm, recipes, and rlang.
  This does not support data columns of type character or date. This function
  also doesn't support columns that only contain missing values. Currently
  supports regression and binomial output.
}
\examples{
library(dplyr)
d <-
  pima_diabetes \%>\%
  mutate_if(is.character, as.factor)


# finds interactions for binomial classification
interactions <- find_interactions(d, diabetes)

# finds interactions for regression classification
interactions <- find_interactions(d, pregnancies)

# Choosing the greedy instead
interactions <- find_interactions(d, pregnancies, brute_force = FALSE)

# Choosing the threshold for the greedy
interactions <- find_interactions(d, pregnancies, brute_force = FALSE,
                                  greedy_cutoff = .02)
}
