% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp_fun.R
\name{find_numeric_columns}
\alias{find_numeric_columns}
\title{Finds numeric columns}
\usage{
find_numeric_columns(d)
}
\arguments{
\item{d}{data frame}
}
\value{
column names that are numeric
}
\description{
Finds numeric columns
}
