% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_unique_columns.R
\name{find_unique_columns}
\alias{find_unique_columns}
\title{Find categorical columns with all unique values and return a list of
columns to ignore.}
\usage{
find_unique_columns(data)
}
\arguments{
\item{data}{A dataframe or tibble containing data to find unique columns.}
}
\value{
A variable containing the names of the categorical columns with
all unique values
}
\description{
\code{find_unique_columns} will find categorical columns in
your data that have all uniuqe values and then return a warning about those
columns and will add those columns to a list of columns to ignore.
}
\examples{
library(dplyr)
d <- data.frame(id_field = c("A","B","C","D"),
test1_field = c(10,20,30,40),
test2_field = c(100,200,300,300),
test3_field = c("A1","B1","B1","D1"),
test4_field = c("AA","BB","CC","DD"))
find_unique_columns(d)
}
