% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{findAltScenarioForEachTestRow}
\alias{findAltScenarioForEachTestRow}
\title{For each test row, find alternate scenarios that reduce probability}
\usage{
findAltScenarioForEachTestRow(trainingData, testData, originalPredictions,
  SDChange, typeOfModel, modelObj, colsToAlter,
  removeIncreasedProbRows = TRUE)
}
\arguments{
\item{trainingData}{Data frame used for training model. This is needed for 
the standard deviation calculations.}

\item{testData}{Data frame. Reasonable alternate scenarios are calculated
for each row.}

\item{SDChange}{Number used to defines what a reasonable person change is.}

\item{typeOfModel}{String. Can Be 'lasso' 'rf' or 'lmm'.}

\item{modelObj}{The actual model object.}
}
\description{
For each test row, drop/add by some SD fraction to see if there
are plausible feature changes that lower the probability of something
}
\examples{
library(caret)
df <- data.frame(a=c(1,2,3,1,1),
                 b=c('m','f','f','m','m'),
                 c=c(0.7,1.4,2.4,2.0,3.4),
                 d=c(100,250,200,150,100),
                 y=c('n','n','y','y','n'))

train <- df[1:3,]
test <- df[4:5,]
test$y <- NULL

glmOb <- glm(y ~ a + b + c + d,
             data = train,
             family = binomial(link = "logit"))

originalPred <- predict(object = glmOb,
                        newdata = test,
                        type = 'response')
originalPred

resultAltScen <- findAltScenarioForEachTestRow(trainingData=train,
                                               testData=test,
                                               originalPredictions=originalPred,
                                               SDChange=0.5,
                                               typeOfModel='lasso',
                                               modelObj=glmOb,
                                               colsToAlter=c('a','c','d'),
                                               removeIncreasedProbRows=FALSE) 

resultAltScen
}
\references{
\url{http://healthcare.ai}
}
\seealso{
\code{\link{healthcareai}}
}

