% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{findBestAlternateScenarios}
\alias{findBestAlternateScenarios}
\title{Find most biggest drop in predictive probability across alternate features}
\usage{
findBestAlternateScenarios(dfAlternateFeat, originalRow, predictionVector,
  predictionOriginal, removeIncreasedProbRows = TRUE)
}
\arguments{
\item{dfAlternateFeat}{Data frame of alternate feature values}

\item{originalRow}{Row from original data frame upon which alternates are
based}

\item{predictionVector}{List of alternate predictions}

\item{predictionOriginal}{Scalar representing original prediction for
row, without alternative scenario}
}
\description{
Compare each alternate probability prediction and determine
which ones are lowest compared to the original; return the top three column
names that lead to the biggest drop and their target value.
}
\examples{
library(caret)
X <- data.frame(a=c(1,2,3,1),
                b=c('m','f','f','m'),
                c=c(0.7,1.4,2.4,2.0),
                d=c(100,250,200,150))

y <- c('y','n','y','n')

X_alt <- calculateSDChanges(df = X,
                            rowToBeAltered = X[4,],
                            sizeOfSDPerturb = 0.5,
                            colsToAlter = c('a','c'))

combined = cbind(X,y)

glmOb <- glm(y ~ a + b + c + d,
             data = combined,
             family = binomial(link = "logit"))

originalPred <- predict(object = glmOb,
                        newdata = X[4,],
                        type = 'response')

alternatePred <- calulcateAlternatePredictions(df = X_alt,
                                               modelObj = glmOb,
                                               type = 'lasso',
                                               removeCols = 'AlteredCol')

dfResult <- findBestAlternateScenarios(dfAlternateFeat = X_alt,
                                       originalRow = X[4,],
                                       predictionVector = alternatePred,
                                       predictionOriginal = originalPred)

dfResult
}
\references{
\url{http://healthcare.ai}
}
\seealso{
\code{\link{healthcareai}}
}

