% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_models.R
\name{flash_models}
\alias{flash_models}
\title{Train models without tuning for performance}
\usage{
flash_models(d, outcome, model_class, models = c("rf", "knn"),
  hyperparameters, metric)
}
\arguments{
\item{d}{A data frame}

\item{outcome}{Name of the column to predict}

\item{model_class}{"regression" or "classification". If not provided, this
will be determined by the class of `outcome` with the determination
displayed in a message.}

\item{models}{Names of models to try, by default "rf" for random forest and
"knn" for k-nearest neighbors. See \code{\link{supported_models}} for
available models.}

\item{hyperparameters}{Optional list of hyperparameters to use. If missing,
default values will be used. If provided, must be a named list of named
lists where the outer list contains models and the inner lists contain
hyperparameter values. See
\code{healthcareai:::get_hyperparameter_defaults()} for a template.}

\item{metric}{What metric to use to assess model performance? Options for
regression: "RMSE" (root-mean-squared error, default), "MAE" (mean-absolute
error), or "Rsquared." For classification: "ROC" (area under the receiver
operating characteristic curve), or "PR" (area under the precision-recall
curve).}
}
\value{
A model_list object
}
\description{
Train models without tuning for performance
}
\details{
This function has two major differences from
  \code{\link{tune_models}}: \enumerate{\item{It uses default hyperparameter
  values to train models instead of using cross-validation to optimize
  hyperparameter values for predictive performance.} \item{It is much
  faster.}}
}
\seealso{
\code{\link{tune_models}}, \code{\link{prep_data}},
  \code{\link{predict.model_list}}, \code{\link{supported_models}}
}
