% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateAUC.R
\name{generateAUC}
\alias{generateAUC}
\title{Generate ROC curve for a dataset.}
\usage{
generateAUC(predictions, labels, aucTYPE, plotFLG)
}
\arguments{
\item{predictions}{A vector of predictions from a machine learning model.}

\item{labels}{A vector of the true labels. Must be the same length as 
predictions.}

\item{aucTYPE}{Can be "SS" or "PR". Defaults to SS.}

\item{plotFLG}{Binary value controlling plots Defaults to TRUE (yes).}
}
\value{
auc Integral AUC of ROC plot.
}
\description{
Generates ROC curve and AUC for Sensitivity/Specificity or 
Precision/Recall.
}
\examples{

}
\references{
\url{http://healthcare.ai}
}
\seealso{
\code{\link{healthcareai}}
}

