% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_hyperparameters.R
\name{get_hyperparameter_defaults}
\alias{get_hyperparameter_defaults}
\title{Get default hyperparameter lists that can be passed to flash_models}
\usage{
get_hyperparameter_defaults(models = get_supported_models(), n = 100,
  k = 10, model_class = "classification")
}
\arguments{
\item{models}{which algorithms?}

\item{n}{Number observations}

\item{k}{Number features}

\item{model_class}{"classification" or "regression"}
}
\value{
List of named data frames that can be passed to
  \code{\link{flash_models}}
}
\description{
Get default hyperparameter lists that can be passed to flash_models
}
\details{
Get default hyperparameters to use, e.g. in flash_models. Each data
  frame in the list corresponds to an algorithm, and each column in each data
  fram corresponds to a hyperparameter for that algorithm.

  k-NN defaults are from the kknn package: kmax = 7, distance = 2
  (Minkowski's exponent, i.e. Euclidean distance), kernal = "optimal"

  Random forest defaults are from Intro to Statistical Learning and caret:
  mtry = sqrt(k), splitrule = "extratrees", min.node.size = 1 for
  classification, 5 for regression
}
