% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantitative_interactions.R
\name{get_interactions}
\alias{get_interactions}
\title{User function}
\usage{
get_interactions(d, outcome)
}
\arguments{
\item{d}{A data frame from \code{\link{prep_data}}. If you want to prepare
your data on your own, use \code{prep_data(..., no_prep = TRUE)}.}

\item{outcome}{Name of the column to predict. It must match the
outcome provided to \code{\link{prep_data}}.}
}
\description{
User function
}
