% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_thresholds.R
\name{get_thresholds}
\alias{get_thresholds}
\title{Get class-separating thresholds for classification predictions}
\usage{
get_thresholds(x, measures = c("cost", "acc", "tpr", "fnr", "tnr", "fpr",
  "ppv", "npv"), cost.fp = 1, cost.fn = 1)
}
\arguments{
\item{x}{Either a predictions data frame (from \code{predict}) or a
model_list (e.g. from \code{machine_learn}).}

\item{measures}{Character vector of performance metrics to calculate. The
returned data frame will have one column for each metric. Any
  of the non-scaler measure arguements to \code{ROCR::performance} should work.
  Defaults to all of the following: \itemize{
  \item{cost: Captures how bad all the errors are. You can adjust the relative costs
   of false alarms and missed detections by setting \code{cost.fp} or
   \code{cost.fn}}. At the default of equal costs, this is directly inversely
   proportional to accuracy.
  \item{acc: Accuracy}
  \item{tpr: True positive rate, aka sensitivity, aka recall}
  \item{tnr: True negative rate, aka specificity}
  \item{fpr: False positive rate, aka fallout}
  \item{fnr: False negative rate}
  \item{ppv: Positive predictive value, aka precision}
  \item{npv: Negative predictive value}
  }}

\item{cost.fp}{Cost of a false positive. Default = 1. Only affects cost.}

\item{cost.fn}{Cost of a false negative. Default = 1. Only affects cost.}
}
\value{
Tibble with rows for each possible threshold value
and columns for the thresholds and each value in \code{measures}. Rows are
sorted increasing by the first value in measures (cost by default).
}
\description{
Get class-separating thresholds for classification predictions
}
\examples{
machine_learn(pima_diabetes[1:20, ], patient_id, outcome = diabetes,
              models = "rf", tune = FALSE) \%>\%
  get_thresholds()
}
