% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_importance.R
\name{get_variable_importance}
\alias{get_variable_importance}
\title{Get variable importance}
\usage{
get_variable_importance(x)
}
\arguments{
\item{x}{model_list}
}
\value{
Data frame of variables and importance scores
}
\description{
Get variable importance
}
\examples{
m <- machine_learn(pima_diabetes[1:50, ], patient_id, outcome = diabetes)
get_variable_importance(m)
}
