% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{getCutOffs}
\alias{getCutOffs}
\title{Function to return FPR given TPR or recall given precision when an ROC or PR 
curve is created.}
\usage{
getCutOffs(perf, aucType = "SS", allCutoffsFlg = FALSE)
}
\arguments{
\item{df}{A data frame to use with the new supervised model.}
}
\value{
Supervised Model Development Params class
}
\description{
Initialize and populate SupervisedModelDevelopmentParams
}
\references{
\url{http://healthcare.ai}
}
\seealso{
\code{\link{healthcareai}}
}

