% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-LIME.R
\name{getLIMEExplanations}
\alias{getLIMEExplanations}
\title{Use LIME explainer to explain new classification predictions}
\usage{
getLIMEExplanations(explainer, testSet, numLabels, numFeatures = 3,
  grainCol = NULL, predictedCol = NULL)
}
\arguments{
\item{explainer}{A LIME explainer built using buildLIMEExplainer}

\item{testSet}{New data on which predictions have been made}

\item{numLabels}{The number of response class predictions to explain, 
ordered from highest to lowest probability (set to 1 or 2 for binary 
classification, more for multiclass)}

\item{numFeatures}{The number of features to use when explaining a prediction}
}
\description{
Use a LIME explainer built using buildLIMEExplainer to explain 
new predictions
}
