% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-find-variation.R
\name{getPipedValue}
\alias{getPipedValue}
\title{Grab number after single pipe in pipe-delimited string}
\usage{
getPipedValue(string)
}
\arguments{
\item{string}{A string with a pipe}
}
\value{
A number from the original string
}
\description{
For a given string with a pipe, return the number that comes after the pipe
}
\examples{
res <- getPipedValue('hello|23')
res
}
\seealso{
\code{\link{healthcareai}} \code{\link{findVariation}} 
\code{\link{createVarianceTallTable}}
}
