% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-local-interpretability.R
\name{getScaledCoeffs}
\alias{getScaledCoeffs}
\title{Extract an ordered list of coefficients from a linear model}
\usage{
getScaledCoeffs(linearModel, info, orderByMagnitude = FALSE)
}
\arguments{
\item{linearModel}{The linear model whose coefficients you want to extract}

\item{info}{a list indexed containing the standard deviations of all of the
numeric variables}

\item{orderByMagnitude}{A boolean, determining whether or not to order the 
coefficients by magnitude rather than by value}
}
\value{
A vector containing the ordered coefficients of the linear model
}
\description{
Extract an ordered list of coefficients from a linear model in 
decreasing order (by value or by magnitude)
}
\references{
\url{http://healthcare.ai}
}
\seealso{
\code{\link{healthcareai}}
}
