% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-kmeans.R
\name{getSilhouetteInf}
\alias{getSilhouetteInf}
\title{Extract Silhouette Information from Clustering}
\usage{
getSilhouetteInf(x, dist)
}
\arguments{
\item{x}{A vector of clusters, the clustering result returned by kmeans(), 
such as kmeans.fit$clusTer}

\item{dist}{A dissimilarity object resurned by dist()}
}
\value{
An n x 3 matrix with columns of cluster, neighbor and silhouette width
}
\description{
Compute silhouette information according to a given k means
clustering result
}
\examples{
data(iris)
head(iris)
kmeans.fit <- kmeans(scale(iris[,1:4]),3)
cluster <- kmeans.fit$cluster
dis <- dist(scale(iris[,1:4]))
res <- getSilhouetteInf(cluster, dis)
str(res)

}
\references{
\url{http://healthcare.ai}

silhouette() in package cluster
\url{https://github.com/pimentel/cluster/blob/master/R/silhouette.R}
}
\seealso{
\code{\link{healthcareai}}
}
