% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_train_test.R
\name{group_strat_split}
\alias{group_strat_split}
\title{Split groups in data training and test data frames}
\usage{
group_strat_split(d, outcome, percent_train = 0.8, grouping_col,
  aggreg_func = dplyr::first)
}
\arguments{
\item{d}{Data frame}

\item{outcome}{Target column, unquoted. Split will be stratified across this
variable}

\item{percent_train}{Proportion of rows in d to put into training. Default is 0.8}

\item{grouping_col}{column name that specifies grouping. Individuals in the same
group are in the same training/test set.}

\item{aggreg_func}{when the data is grouped, the outcome column is aggregated
based on this function.}
}
\value{
A list of two data frames with names train and test
}
\description{
Split groups in data training and test data frames
}
