% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{GroupedLOCF}
\alias{GroupedLOCF}
\title{Last observation carried forward}
\usage{
GroupedLOCF(df, id)
}
\arguments{
\item{df}{data frame sorted by an ID column and a time or sequence number
column.}

\item{id}{A column name (in ticks) in df to group rows by.}
}
\value{
A data frame where the last non-NA values are carried forward
(overwriting NAs) until the group ID changes.
}
\description{
Carries the last observed value forward for all columns in a data.table
grouped by an id.
}
\examples{
library(HCRTools)
df = data.frame(PersonID=c(1,1,2,2,3,3,3),
                wt=c(.5,NA,NA,NA,.3,.7,NA),
                ht=c(NA,1,3,NA,4,NA,NA),
                date=c('01/01/2015','01/15/2015','01/01/2015','01/15/2015',
                       '01/01/2015','01/15/2015','01/30/2015'))

head(df,n=7)

df_result = GroupedLOCF(df, 'PersonID')

head(df_result, n = 7)
}
\references{
\url{https://community.healthcatalyst.com/community/data-science}
}
\seealso{
\code{\link{HCRTools}}
}

