% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-testing-utils.R
\name{ignoreSpecWarn}
\alias{ignoreSpecWarn}
\title{Function to suppress specific warnings in unit tests}
\usage{
ignoreSpecWarn(code, wRegexps)
}
\arguments{
\item{code}{A piece of code to run, for which the warnings should be 
suppressed}

\item{wRegexps}{A vector of regular expressions corresponding to the 
warnings that should be suppressed}
}
\description{
This function allows one to suppress one or more specific 
warnings.  The intended purpose is to suppress warning messages that are 
expected when running specific unit tests, but are unrelated to the 
functionality being tested.
}
\references{
\url{http://healthcareai-r.readthedocs.io}
}
\seealso{
\code{\link{healthcareai}}
}
