% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{ImputeColumn}
\alias{ImputeColumn}
\title{Perform imputation on a vector}
\usage{
ImputeColumn(v)
}
\arguments{
\item{v}{A vector, or column of values with NAs.}
}
\value{
A vector, or column of values now with no NAs
}
\description{
This class performs imputation on a vector. For numeric vectors
the vector-mean is used; for factor columns, the most frequent value is used.
}
\examples{
# For a numeric vector
v_result = ImputeColumn(c(1,2,3,NA))

# For a factor vector
v_result = ImputeColumn(c('Y','N','Y',NA))

# To use this function on an entire dataframe:
df = data.frame(a=c(1,2,3,NA),
                b=c('Y','N','Y',NA))
df[] <- lapply(df, ImputeColumn)
}
\seealso{
\code{\link{HCRTools}}
}

