% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-testing-utils.R
\name{initializeParamsForTesting}
\alias{initializeParamsForTesting}
\title{Function to initialize and populate the SupervisedModelDevelopmentParams each 
time a unit test is run.}
\usage{
initializeParamsForTesting(df)
}
\arguments{
\item{df}{A data frame to use with the new supervised model.}
}
\value{
Supervised Model Development Params class
}
\description{
Initialize and populate SupervisedModelDevelopmentParams
}
\references{
\url{http://healthcareai-r.readthedocs.io}
}
\seealso{
\code{\link{healthcareai}}
}
