% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.R
\name{interpret}
\alias{interpret}
\title{Interpret a model via regularized coefficient estimates}
\usage{
interpret(x, sparsity = NULL)
}
\arguments{
\item{x}{a model_list object containing a glmnet mdoel}

\item{sparsity}{If NULL (default) coefficients for the best-performing model
will be returned. Otherwise, a value in [0, 1] that determines the
sparseness of the model for which coefficients will be returned, with 0
being maximally sparse (i.e. having the fewest non-zero coefficients) and 1
being minimally sparse.}
}
\value{
A data frame of variables and their regularized regression
  coefficient estimates.
}
\description{
Interpret a model via regularized coefficient estimates
}
\details{
If x was trained with more than one value of alpha the best value of
  alpha is used; sparsity is determined only via the selection of lambda.

  Coefficients are on the scale of the predictors; they are not standardized,
  so unless features were scaled before training (e.g. with
  \code{prep_data(..., scale = TRUE)}, the magnitude of coefficients does not
  necessarily reflect their importance.
}
\examples{
m <- machine_learn(pima_diabetes, patient_id, outcome = age, models = "glm")
interpret(m)
interpret(m, .2)
}
