% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_list_generics.R
\name{is.model_list}
\alias{is.model_list}
\alias{is.classification_list}
\alias{is.regression_list}
\alias{is.multiclass_list}
\title{Type checks}
\usage{
is.model_list(x)

is.classification_list(x)

is.regression_list(x)

is.multiclass_list(x)
}
\arguments{
\item{x}{Object}
}
\value{
Logical
}
\description{
Type checks
}
