% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{IsBinary}
\alias{IsBinary}
\title{Check if a vector has only two unique values.}
\usage{
IsBinary(v)
}
\arguments{
\item{v}{A vector, or column of values}
}
\value{
A boolean
}
\description{
Check if a vector is binary (not counting NA's)
}
\examples{
IsBinary(c(1,2,NA))
IsBinary(c(1,2,3))
}
\references{
\url{http://healthcareml.org/}
}
\seealso{
\code{\link{HCRTools}}
}

