% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-column-row-expectations.R
\name{isCategorical}
\alias{isCategorical}
\title{Check if a data frame only has categorical columns.}
\usage{
isCategorical(df, maxLevels)
}
\arguments{
\item{df}{A dataframe}
}
\value{
A boolean
}
\description{
Check if a dataframe only has categorical columns
}
\examples{

}
\references{
\url{http://healthcare.ai}
}
\seealso{
\code{\link{healthcareai}}
}
