% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-error-checking-utils.R
\name{isnotYNOrNA}
\alias{isnotYNOrNA}
\title{Boolean. Tests whether predictedCol is Y/N}
\usage{
isnotYNOrNA(x)
}
\arguments{
\item{x}{A data frame column, vector, or matrix column}
}
\value{
TRUE or FALSE depending on test
}
\description{
Returns a logical, TRUE or FALSE, depending on what is
contained in the df or vector
}
\examples{
dat <- data.frame(a = c(0,1,1,0,0,0,1,1,1,0,1,0,1,0,1,0,1,1,1,0))
dat2 <- data.frame(a = c(3, 4, 5, 6, 7, 8, 9))
dat3 <- data.frame(a = c('Y', 'N', 'Y', 'N'))
isnotYN(dat[, 1])
isnotYN(dat2[, 1]) 
isnotYN(dat3[, 1])              
                
}
\references{
\url{http://healthcare.ai}
}
\seealso{
\code{\link{healthcareai}}
}
