% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-column-row-expectations.R
\name{isNumeric}
\alias{isNumeric}
\title{Check if a data frame only has numeric columns.}
\usage{
isNumeric(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
A boolean
}
\description{
Check if a dataframe only has numeric columns
}
\examples{

}
\references{
\url{http://healthcare.ai}
}
\seealso{
\code{\link{healthcareai}}
}
