% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-feature-availability-profiler.R
\name{isZeroOne}
\alias{isZeroOne}
\title{Boolean. Finds any 0 or 1 in predicted col. Works on factor, character,
numeric, and integer variables}
\usage{
isZeroOne(x)
}
\arguments{
\item{x}{A data frame column, vector, or matrix column}
}
\value{
TRUE or FALSE depending on test
}
\description{
Returns a logical, TRUE or FALSE, depending on what is
contained in the df or vector
}
\examples{
dat <- data.frame(a = c(0,1,1,0,0,0,1,1,1,0,1,0,1,0,1,0,1,1,1,0))
dat2 <- data.frame(a = c(3, 4, 5, 6, 7, 8, 9))
isZeroOne(dat[, 1])
isZeroOne(dat2[, 1])               
                
}
\references{
\url{http://healthcare.ai}
}
\seealso{
\code{\link{healthcareai}}
}
