% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans-clustering-method.R
\docType{class}
\name{KmeansClustering}
\alias{KmeansClustering}
\title{Build clusters using kmeans()}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
...
}
\arguments{
\item{object}{of UnsuperviseModelParameters class for $new() constructor}

\item{dataType}{The type of your data set('numeric' for all numeric data set,
'categorical' for all categorical data set, and "mixed" for data set that has both
numeric and categorical variables)}

\item{df}{Dataframe whose columns are used for calc.}

\item{grainCol}{Optional. The dataframe's column that has IDs pertaining to 
the grain. No ID columns are truly needed for this step.}

\item{labelCol}{Optional.}

\item{numOfCluster}{Number of clusters you want to build.}

\item{impute}{Set all-column imputation to F or T.
This uses mean replacement for numeric columns
and most frequent for factorized columns.
F leads to removal of rows containing NULLs.}

\item{debug}{Provides the user extended output to the console, in order
to monitor the calculations throughout. Use T or F.}
}
\description{
This step allows you to use kmeans() to build clusters, based on
your data.
}
\examples{

}
\references{
\url{http://hctools.org/}
}
\seealso{
\code{\link{LassoDevelopment}}

\code{\link{LinearMixedModelDevelopment}}

\code{\link{healthcareai}}
}
\keyword{datasets}
