% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearMixedModel.R
\docType{class}
\name{LinearMixedModel}
\alias{LinearMixedModel}
\title{Compare predictive models, created on your data}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LinearMixedModel
}
\arguments{
\item{object}{of SuperviseModelParameters class for $new() constructor}

\item{type}{The type of model (either 'regression' or 'classification')}

\item{df}{Dataframe whose columns are used for calc.}

\item{grainCol}{The data frame's ID column pertaining to the grain}

\item{personCol}{The data frame's ID column pertaining to the person/patient}

\item{predictedCol}{Column that you want to predict.}

\item{impute}{Set all-column imputation to F or T.
This uses mean replacement for numeric columns
and most frequent for factorized columns.
F leads to removal of rows containing NULLs.}

\item{debug}{Provides the user extended output to the console, in order
to monitor the calculations throughout. Use T or F.}
}
\description{
This step allows one to create test models on your data
and helps determine which performs best.
}
\examples{

### Doing classification
library(HCRTools)
library(lme4)

df <- sleepstudy

str(df)

# Create binary column for classification
df$ReactionFLG <- ifelse(df$Reaction > 300, 'Y','N')
df$Reaction <- NULL

set.seed(42)
p <- SupervisedModelParameters$new()
p$df = df
p$type = 'classification'
p$impute = TRUE
p$personCol = 'Subject' # Think of this as PatientID
p$predictedCol = 'ReactionFLG'
p$debug = TRUE
p$cores = 1

# Create Mixed Model
lmm <- LinearMixedModel$new(p)
lmm$run()

### Doing regression
library(HCRTools)
library(lme4)

# SQL query and connection goes here - see SelectData function.

df <- sleepstudy

# Add GrainID, which is equivalent to PatientEncounterID
df$GrainID <- seq.int(nrow(df))

str(df)

set.seed(42)
p <- SupervisedModelParameters$new()
p$df = df
p$type = 'regression'
p$impute = TRUE
p$grainCol = 'GrainID' # Think of this as PatientEnounterID
p$personCol = 'Subject' # Think of this as PatientID
p$predictedCol = 'Reaction'
p$debug = FALSE
p$cores = 1

# Create Mixed Model
lmm <- LinearMixedModel$new(p)
lmm$run()

}
\references{
\url{http://healthcareml.org/}
}
\seealso{
\code{\link{HCRTools}}
}
\keyword{datasets}

