% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-kmeans.R
\name{lineMagnitude}
\alias{lineMagnitude}
\title{Compute the distance between two points}
\usage{
lineMagnitude(x1, y1, x2, y2)
}
\arguments{
\item{x1}{A number, x axis of one point}

\item{y1}{A number, y axis of one point}

\item{x2}{A number, x axis of the other point}

\item{y2}{A number, y axis of the other point}
}
\value{
A number, the length between two points.
}
\description{
compute the distance between two points given the x, y axes of the
points.
}
\examples{
lineMagnitude(1,2,5,3)
}
\references{
\url{http://healthcare.ai}
}
\seealso{
\code{\link{healthcareai}}
}
