% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_variables.R
\name{list_variables}
\alias{list_variables}
\title{Takes a vector of variables and returns the unique variable in a string}
\usage{
list_variables(my_variables)
}
\arguments{
\item{my_variables}{a vector of categorical variables}
}
\value{
a string that lists all the unique elements of the vector in sentence form
}
\description{
Takes a vector of variables and returns the unique variable in a string
}
