% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-local-interpretability.R
\name{localPerturbations}
\alias{localPerturbations}
\title{Get a dataframe consisting of rows "near" a given row}
\usage{
localPerturbations(baseRow, modifiableCols, info, size = 1000, spread = 1/4,
  grainCol = NA, predictedCol = NA)
}
\arguments{
\item{baseRow}{a row of data (in a dataframe) to perturb}

\item{modifiableCols}{a vector of column names corresponding to the columns
which should be perturbed}

\item{info}{a list which indexed by the column names in modifiableCols 
containing the standard deviation (for numeric columns) or the factor levels
(for factor columns)}

\item{size}{the number of rows in}

\item{spread}{a number representing the size of the perturbations}

\item{grainCol}{the name of the grain column in baseRow, if present}

\item{predictedCol}{the name of the predicted column in baseRow, if present}
}
\value{
a dataframe consisting whose rows are perturbations of baseRow
}
\description{
Build a dataframe consisting of rows which are perturbations 
of a given row.
}
\references{
\url{http://healthcare.ai}
}
\seealso{
\code{\link{healthcareai}}
}
