% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingness.R
\name{make_na}
\alias{make_na}
\title{Replaces missing data with NA}
\usage{
make_na(d, to_replace, drop_levels = TRUE)
}
\arguments{
\item{d}{A dataframe or tibble}

\item{to_replace}{A vector of the values or string of the value that will be
replaced with NA}

\item{drop_levels}{a boolean to drop unused levels in factor vectors that are
created from missingness values being replaced. Default = TRUE}
}
\value{
A tibble with the missing data replaced with NA
}
\description{
This function replaces any given value in a dataframe or tibble
  with NA. It can be used when missing values are filled with characters like
  "missing". Numeric vector columns that were originally loaded as character
  or factor columns (because of a missingness string), are converted to
  numeric vectors when the missingness string is replaced.
}
\examples{
# Replace "missing" in dat
dat <- data.frame(gender = c("male", "male", "female", "male", "missing"),
              name = c("Paul", "Jim", "Sarah", "missing", "Alex"),
              weight = c(139, 0, 193, 158, 273))

make_na(dat, "missing")

# Replace "missing" in dat. `gender` and `name` factor vectors will keep
# "missing" as a level
dat <- data.frame(gender = c("male", "male", "female", "male", "missing"),
              name = c("Paul", "Jim", "Sarah", "missing", "Alex"),
              weight = c(139, 0, 193, 158, 273))

make_na(dat, "missing", drop_levels = FALSE)

# Replace both "missing" and "??" in dat. `age` should be loaded as a numeric
# vector, but since "missing" is present it is stored as a character vector.
# `age` will be converted to a numeric vector when "missing" is replaced.
dat <- data.frame(gender = c("male", "??", "female", "male", "missing"),
              age = c(64, 52, 75, "missing", 70),
              weight = c(139, 0, 193, "??", 273),
              stringsAsFactors = FALSE)

make_na(dat, c("missing", "??"))

}
