% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Mode}
\alias{Mode}
\title{Mode}
\usage{
Mode(x)
}
\arguments{
\item{x}{Either a vector or a frequency table from \code{table}}
}
\value{
The modal value of x
}
\description{
Mode
}
\examples{
x <- c(3, 1:5)
Mode(x)
Mode(table(x))
}
