% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-local-interpretability.R
\name{modifiableFactors1Row}
\alias{modifiableFactors1Row}
\title{Build a dataframe with the modifiable factors information for 1 row}
\usage{
modifiableFactors1Row(baseRow, modifiableCols, info2, predictFunction,
  lowerProbGoal = TRUE, numberOfPercentiles = 12)
}
\arguments{
\item{baseRow}{the row of data for which to determine the modifiable factors}

\item{modifiableCols}{a list of names of the modifiable columns}

\item{info2}{percentiles and frequencies TODO: expand}

\item{predictFunction}{a function used to make new predictions using the 
model we wish to study}

\item{lowerProbGoal}{a boolean specifying whether the goal is to reduce the
predicted probability (i.e., Y corresponds to negative outcome)}

\item{numberOfPercentiles}{the number of percentiles to use when
approximating the model}
}
\value{
a dataframe whose rows all agree except in the \code{variable}
column which takes each level in \code{factorLevels} exactly once.
}
\description{
Build a dataframe with the modifiable factors information for 1 
row. This dataframe contains the
}
\references{
\url{http://healthcareai-r.readthedocs.io}
}
\seealso{
\code{\link{healthcareai}}
}
