% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-nelson-rules.R
\name{nelsonRule2}
\alias{nelsonRule2}
\title{Analyze points in time to determine whether or not Nelson Rule 2 was violated}
\usage{
nelsonRule2(df, measure_col, date_col, plot_flg = TRUE)
}
\arguments{
\item{df}{A data frame}

\item{measure_col}{A string denoting the column that contains the numeric
value to be evaluated by the nelsonRule2 function}

\item{date_col}{A string denoting the date column to be evaluated by the 
nelsonRule2 function}

\item{plot_flg}{A binary indicator of whether or not to include a plot in the
output of the function}
}
\value{
A data frame containing the date, measure value, a flag indicating 
whether or not a rule violation occured and a description of the violation.
}
\description{
Search within a data frame to find violations of Nelson Rule 2.
Nelson Rule 2: Nine or more points in a row are on the same side of the mean.
}
\examples{
library(healthcareai)
date <- seq.Date(from = as.Date('2016-01-03'), length.out = 104, by = 'week')
set.seed(34)
measureValue <- rnorm(length(date), mean = 100, sd = 15)
d <- data.frame(date, measureValue)

# Alter some measureValues so a violation occurs
d[['measureValue']][33:41] <- runif(n = 9,min = 105, max = 135)
d[['measureValue']][78:87] <- runif(n = 10,min = 80, max = 95)

nelsonRule2(d, 'measureValue', 'date')
}
\references{
\url{http://healthcareai-r.readthedocs.io}
\url{https://en.wikipedia.org/wiki/Nelson_rules}
}
\seealso{
\code{\link{healthcareai}}
}
