% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{OrderByDate}
\alias{OrderByDate}
\title{Order the rows in a data frame by date}
\usage{
OrderByDate(df, date_col, descending = FALSE)
}
\arguments{
\item{df}{A data frame}

\item{date_col}{Name of column in data frame that contains dates}

\item{descending}{Boolean for whether the output should be in descending order}
}
\value{
A data frame ordered by date column
}
\description{
Returns a data frame that's ordered by its date column
}
\examples{
library(HCRTools)
df = data.frame(date=c('2009-01-01','2010-01-01','2009-03-08','2009-01-19'),
                a=c(1,2,3,4))
df_result = OrderByDate(df,'date', descending=FALSE)
head(df_result)
}
\references{
\url{http://healthcareml.org/}
}
\seealso{
\code{\link{HCRTools}}
}

