% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-kmeans.R
\name{pcaAnalysis}
\alias{pcaAnalysis}
\title{Perform principle component analysis}
\usage{
pcaAnalysis(df)
}
\arguments{
\item{df}{A numeric data frame w/o NA's}
}
\value{
A list that contains the data frame of principle components and 
the proportion of variance explained by each PC.
}
\description{
performs PCA on numeric data frames
}
\examples{
data(iris)
head(iris)
df <- iris[,1:4]
res <- pcaAnalysis(df)
head(res[[1]])
}
\references{
\url{http://healthcare.ai}
}
\seealso{
\code{\link{healthcareai}}
}
