% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{percentDataAvailableInDateRange}
\alias{percentDataAvailableInDateRange}
\title{Find the percent of a column that's filled}
\usage{
percentDataAvailableInDateRange(df, dateColumn = NULL,
  startInclusive = NULL, endExclusive = NULL)
}
\arguments{
\item{df}{A dataframe
filled by hour are the columns}

\item{keyList}{A vector of strings, representing the features.}
}
\value{
Nothing
}
\description{
Shows what percentage of data is avilable (potentially within a specified 
date range)
}
\references{
\url{http://healthcare.ai}
}
\seealso{
\code{\link{healthcareai}}
}
