% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pip.R
\name{permute_process_variables}
\alias{permute_process_variables}
\title{Take a dataframe and build a larger dataframe by permuting the
values in certain columns.}
\usage{
permute_process_variables(dataframe, variable_levels, variable_direction)
}
\arguments{
\item{dataframe}{A dataframe.}

\item{variable_levels}{A list indexed by the modifiable
process variables and containing the factor levels of each such variable.}
}
\value{
A large dataframe where the values in the modifiable variable
columns have been permuted.
}
\description{
Take a dataframe and build a larger dataframe by permuting the
values in the columns corresponding to \code{variable_levels}.
}
\keyword{internal}
