% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pima_diabetes}
\alias{pima_diabetes}
\title{Patient diabetes dataset}
\format{A tibble data frame with 768 rows and 10 variables:}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/pima+indians+diabetes}
}
\usage{
pima_diabetes
}
\description{
A dataset containing diabetes status and other health-related variables
for 768 females, at least 21 years old, of Pima Indian heritage.
}
\keyword{datasets}
