% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_plot.R
\name{plot.cf_df}
\alias{plot.cf_df}
\title{Plot Simulated Counterfactual Predictions}
\usage{
\method{plot}{cf_df}(x, reorder_categories = TRUE, n_use = 4,
  aggregate_fun = median, jitter_y = TRUE, x_var, color_var,
  font_size = 11, strip_font_size = 0.85, line_width = 0.5,
  line_alpha = 0.7, rotate_x = FALSE, print = TRUE, ...)
}
\arguments{
\item{x}{A cf_df object from \code{link{simulate}}}

\item{reorder_categories}{If TRUE (default) categorical variables that were
varied in \code{simulate} will be arranged in decreasing order of their
median predicted outcome}

\item{aggregate_fun}{Default = median. Varying features in x are mapped to
the x-axis, line color, and vertical- and horizontal facets. If
more than four features vary, this function is used to aggreagate across
the least-important varying features.}

\item{jitter_y}{If TRUE (default) and a variable is mapped to color (i.e. if
there is more than one varying variable), the vertical location of the
lines will be jittered slightly (no more than 1% of the range of the
outcome variable) to avoid overlap.}

\item{x_var}{Variable to put on the x-axis (unquoted). If not provided, the
most important variable is used, with numerics prioritized if one was
varied}

\item{color_var}{Variable to color lines (unquoted). If not provided, the
most important variable excluding \code{x_var}}

\item{font_size}{Parent font size for the whole plot. Default = 11}

\item{strip_font_size}{Relative font size for facet strip title font. Default
= 0.85}

\item{line_width}{Width of lines. Default = 0.5}

\item{line_alpha}{Opacity of lines. Default = 0.7}

\item{rotate_x}{If FALSE (default), x axis tick labels are positioned
horizontally. If TRUE, they are rotated one quarter turn, which can be
helpful when a categorical variable with long labels is mapped to x.}

\item{print}{Print the plot? Default is FALSE. Either way, the plot is
invisbly returned}

\item{...}{}
}
\value{
ggplot object, invisibly
}
\details{

}
