% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_predictions.R
\name{plot.hcai_predicted_df}
\alias{plot.hcai_predicted_df}
\title{Plot model predictions vs observed outcomes}
\usage{
\method{plot}{hcai_predicted_df}(x, outcomes = NULL, print = TRUE, ...)
}
\arguments{
\item{x}{data frame as returned `predict.model_list`}

\item{outcomes}{Vector of outcomes if not present in x}

\item{print}{Logical, if TRUE (default) the plot is printed on the current
graphics device. The plot is always (silently) returned.}

\item{...}{Parameters to pass to plot_regression_predictions or
plot_classification_predictions}
}
\value{
A ggplot object
}
\description{
Plot model predictions vs observed outcomes
}
\details{
The following arguments can be provided to customize the plot: For
  regression: title, point_size, point_alpha, font_size. For classification:
  title, fill_colors, fill_alpha, curve_flex, font_size. For details on how
  to use them, see \code{\link{plot_regression_predictions}} or
  \code{\link{plot_classification_predictions}}.
}
\examples{
models <- machine_learn(pima_diabetes[1:50, ], patient_id, outcome = plasma_glucose,
                        models = "rf", tune = FALSE)
predictions <- predict(models)
plot(predictions)
plot(predictions, title = "This model's predictions regress to the mean",
     point_size = 3, point_alpha = .7, font_size = 14)
}
