% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_list.R
\name{plot.regression_list}
\alias{plot.regression_list}
\title{Plot performance of regression models}
\usage{
plot.regression_list(rlist, print = TRUE)
}
\arguments{
\item{rlist}{regression_list object as returned by \code{\link{tune}}}

\item{print}{If TRUE (default) plot is printed}
}
\value{
Plot of model performance as a function of algorithm and
  hyperparameter values tuned over. Generally called for the side effect of
  printing a plot, but the plot is also invisibly returned.
}
\description{
Plot performance of regression models
}
\examples{
mtcars \%>\%
 tune(mpg) \%>\%
 plot()
}
