% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{plot.simulated_df}
\alias{plot.simulated_df}
\title{Plot Simulated Counterfactual Predictions}
\usage{
\method{plot}{simulated_df}(x, font_size = 11, print = TRUE, ...)
}
\arguments{
\item{x}{A simulated_df object from \code{link{simulate}}}

\item{...}{}
}
\description{
Plot Simulated Counterfactual Predictions
}
