% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_predictions.R
\name{plot_classification_predictions}
\alias{plot_classification_predictions}
\title{Plot predictions from a classification model}
\usage{
plot_classification_predictions(x, title = NULL,
  fill_colors = c("firebrick", "steelblue"), fill_alpha = 0.7,
  curve_flex = 1, font_size = 11, target)
}
\arguments{
\item{x}{hcai_predicted_df from plot.hcai_predicted_df}

\item{title}{Character: Plot title, default NULL produces no title.}

\item{fill_colors}{Length-2 character vector: colors to fill density
curves. Default is c("firebrick", "steelblue"). If named, names must match
\code{unique(x[[target]])}, in any order.}

\item{fill_alpha}{Number in [0, 1] giving opacity of fill colors.}

\item{curve_flex}{Numeric. Kernal adjustment for density curves. Default is 1.
Less than 1 makes curves more flexible, analogous to smaller bins in a
histogram; greater than 1 makes curves more rigid.}

\item{font_size}{Number: Relative size of all font in plot, default = 11}

\item{target}{outcome column name}
}
\value{
ggplot object
}
\description{
Plot predictions from a classification model
}
\details{
This function is not meant to be called directly. Use
  \code{plot(predictions)} instead. You can pass arguments to this function
  (e.g. a plot title) through \code{plot}.
}
