% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantitative_interactions.R
\name{plot_interactions}
\alias{plot_interactions}
\title{plots interactions}
\usage{
plot_interactions(interactions, model, obs, top_n = 6, print = TRUE,
  font_size = 11)
}
\description{
plots interactions
}
