% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_predictions.R
\name{plot_regression_predictions}
\alias{plot_regression_predictions}
\title{Plot predictions from a regression model}
\usage{
plot_regression_predictions(x, title = NULL, point_size = 1,
  point_alpha = 1, font_size = 11, target)
}
\arguments{
\item{x}{hcai_predicted_df from plot.hcai_predicted_df}

\item{title}{Character: Plot title}

\item{point_size}{Number: Point size, relative to 1}

\item{point_alpha}{Number in [0, 1] giving point opacity}

\item{font_size}{Relative size of all font in plot}

\item{target}{outcome column name}
}
\value{
ggplot object
}
\description{
Plot predictions from a regression model
}
\details{
This function is not meant to be called directly. Use
  \code{plot(predictions)} instead. You can pass arguments to this function
  (e.g. a plot title) through \code{plot}.
}
