% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_importance.R
\name{plot_variable_importance}
\alias{plot_variable_importance}
\title{Plot variable importance}
\usage{
plot_variable_importance(x, caption = "model", title = NULL,
  font_size = 11, print = TRUE)
}
\arguments{
\item{x}{Either a model_list object or a data frame from
\code{\link{get_variable_importance}}}

\item{caption}{Either "model", "none", or a string to be used as the plot
caption. "model" puts the name of the best-performing model, on which
variable importances are generated, in the caption.}

\item{title}{Plot title}

\item{font_size}{Relative size for all fonts, default = 11}

\item{print}{Print the plot?}
}
\value{
ggplot object, invisibly
}
\description{
Plot variable importance
}
\examples{
m <- machine_learn(pima_diabetes[1:50, ], patient_id, outcome = diabetes, tune = FALSE)
plot_variable_importance(m)
}
