% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-plot.R
\name{plotPRCurve}
\alias{plotPRCurve}
\title{Plot PR Curves from SupervisedModel classes}
\usage{
plotPRCurve(PRCurves, names, legendLoc)
}
\arguments{
\item{PRCurves}{A vector/array/list of PR curves}

\item{names}{A vector of algorithm/class names}

\item{legendLoc}{Location of the legend string to display}
}
\description{
Plot PRCurves calculated by children classes of SupervisedModel
}
\references{
\url{http://healthcareai-r.readthedocs.io}
}
\seealso{
\code{\link{healthcareai}}
}
