% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-availability-profiler.R
\name{plotProfiler}
\alias{plotProfiler}
\title{Display feature profile over time}
\usage{
plotProfiler(result, keyList)
}
\arguments{
\item{result}{A 2-d list, where features go down the rows and percent
filled by hour are the columns}

\item{keyList}{A vector of strings, representing the features.}
}
\value{
Nothing
}
\description{
Shows what percentage of data is avilable after a particular starting
time period.
}
\references{
\url{http://healthcare.ai}
}
\seealso{
\code{\link{healthcareai}}
}
