% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-plot.R
\name{plotROCs}
\alias{plotROCs}
\title{Plot ROCs from SupervisedModel classes}
\usage{
plotROCs(rocs, names, legendLoc)
}
\arguments{
\item{rocs}{A vector/array/list of ROC values}

\item{names}{A vector of algorithm/class names}

\item{legendLoc}{Location of the legend string to display}
}
\description{
Plot ROCs calculated by children classes of SupervisedModel
}
\references{
\url{http://healthcareai-r.readthedocs.io}
}
\seealso{
\code{\link{healthcareai}}
}
