% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-kmeans.R
\name{plotSilhouette}
\alias{plotSilhouette}
\title{Plot silhouette width}
\usage{
plotSilhouette(x, col = "gray")
}
\arguments{
\item{x}{a matrix returned by getSilhouetteInf()}

\item{col}{colors for different clusters}
}
\value{
a silhouette plot
}
\description{
Plot silhouette width given a matrix that stores the silhouette
information
}
\examples{
data(iris)
head(iris)
kmeans.fit <- kmeans(scale(iris[,1:4]),3)
cluster <- kmeans.fit$cluster
dis <- dist(scale(iris[,1:4]))
res <- getSilhouetteInf(cluster, dis)
plotSilhouette(res,col = c(1:3))

}
\references{
\url{http://healthcare.ai}

plot.silhouette() function in package cluster
\url{https://cran.r-project.org/web/packages/cluster/index.html}
}
\seealso{
\code{\link{healthcareai}}
}
