% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-local-interpretability.R
\name{plotVariableEffects}
\alias{plotVariableEffects}
\title{Plot the output of a model as each coefficient changes individually}
\usage{
plotVariableEffects(baseRow, modifiableCols, info, predictFunction, type,
  spread = 1/2, modifiableDfRow = NULL, grid = NULL)
}
\arguments{
\item{baseRow}{a row in a data frame, used as the reference from which to 
vary the variables}

\item{modifiableCols}{a vector of column names corresponding to the columns
which should be perturbed}

\item{info}{a list which indexed by the column names in modifiableCols 
containing the standard deviation (for numeric columns) or the factor levels
(for factor columns)}

\item{type}{the type of fitObj: "classification" or "regression"}

\item{spread}{a number representing how widely to vary the variables}

\item{modifiableDfRow}{the row in ModifiableFactorsDf corresponding to 
baseRow}

\item{grid}{An ordered pair of integers (rows, columns) which determines the 
number of plots to display at once. The first entry determines the number of
rows and the second determines the number of columns.}

\item{fitObj}{the model for which we are plotting changes in variables}
}
\description{
Plot the output of a model as each coefficient changes 
individually
}
\references{
\url{http://healthcare.ai}
}
\seealso{
\code{\link{healthcareai}}
}
