% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-impact-coding.R
\name{pphcv}
\alias{pphcv}
\title{Pre-Process High-Cardinality Variables}
\usage{
pphcv(df, target, high_card, type, pos_class = NULL, m = NULL,
  multi_base = NULL, remove_high_card = FALSE)
}
\arguments{
\item{df}{A data frame that contains the high-cardinality variable/s and an 
outcome variable of interest.}

\item{target}{The variable/outcome of interest. Enter as a string.}

\item{high_card}{The high-cardinality variable/s. Enter as a string or vector
of strings, i.e. c("variable1", "variable2", "variable3")}

\item{type}{What type of variable is the target? Options are "binary",
"continuous", and "multiclass".}

\item{pos_class}{If target is binary, what is the positive class? Enter as a 
string for a factor or character target and a number for a numeric target.}

\item{m}{A tuneable parameter that depends on the data domain and expected 
noise. If the expected noise is small, m should be small and vice versa. 
Defaults to 1.}

\item{multi_base}{If target is multiclass, multi_base gives the option to 
specify which class to use as the base category. If not specified the first
class will be used as the base level.}

\item{remove_high_card}{If TRUE, removes the original high-cardinality 
variable from the final output dataframe. Defaults to FALSE}
}
\value{
A data frame that contains the new variables consisting of the
probabilities/expectations of the high-cardinality variables. Note: the
original high-cardinality variable/s will be removed from the final output if
specified using remove_high_card.
}
\description{
pphcv stands for pre-processing of high-cardinality variables. 
It is a method to reduce high-cardinality (many categories) variables into 
probabilities of the target class or expectations of the target. This allows 
the user to utilize variables with large numbers of categories and thus 
retain much of the explanatory power of those variables. 

For binary targets, the high-cardinality variables will be reduced to 
probabilities of the target positive class given each high-cardinality level.
For continuous targets, the high-cardinality variables will be reduced to 
expectations of the target variable given each high-cardinality level. For 
multiclass targets with n classes, each high-cardinality variable will be 
transformed into n - 1 attributes representing the probability of each target
class given each level of the high-cardinality variable. 

This technique was taken from a paper titled,"A preprocessing Scheme for High
Cardinality Categorical Attributes in Classification and Prediction 
Problems," written by Daniele Micci-Barreca, published in ACM SIGKDD 
Explorations Newsletter, Volume 3 Issue 1, July 2001, pages 27-32. See
references.
}
\examples{
set.seed(10)
dat <- data.frame(Ycontinuous = sample(5:42, size = 14, replace = TRUE),
                  Ycategorical = sample(c("Y", "N"), size = 14, 
                                       replace = TRUE, prob = c(0.45, 0.55)),
                  YCat01 = c(0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0),
                  hcX = sample(letters[1:5], size = 14, replace = TRUE),
                  hcX2 = sample(letters[6:10], size = 14, replace = TRUE),
                  hcX3 = sample(letters[11:15], size = 14, replace = TRUE),
                  Ymulticlass = sample(c("blue", "green", "red"), size = 14, 
                                       replace = TRUE))
               
# Continuous target and one high_cardinality variable
pphcv(df = dat, target = "Ycontinuous", high_card = "hcX", 
      type = "continuous")
          
# Binary target and one high_cardinality variable
pphcv(df = dat, target = "Ycategorical", high_card = "hcX", pos_class = "Y", 
      type = "binary")
          
# Binary target coded as numeric(0, 1) and one high-cardinality variable 
# and specifying the pos_class.
pphcv(df = dat, target = "YCat01", high_card = "hcX", pos_class = 1, 
      type = "binary")
          
# Binary target coded as numeric(0, 1) and one high-cardinality variable and 
# not specifying the pos_class. Use type = "continuous". Should give same
# results as if type is binary and positive class is specified.
pphcv(df = dat, target = "YCat01",  high_card = "hcX", type = "continuous", 
      m = 1)

# Multiclass target and one high-cardinality variable
pphcv(df = dat, target = "Ymulticlass", high_card = "hcX", 
      type = "multiclass")
          
# Binary target and a vector of high-cardinality variables.
vec <- c("hcX", "hcX2", "hcX3")
pphcv(df = dat, target = "YCategorical", high_card = vec, 
          type = "binary", pos_class = "Y")

# Multiclass target and a vector of high-cardinality variables and specify
# remove_high_card = TRUE
vec <- c("hcX", "hcX2", "hcX3")
pphcv(df = dat, target = "Ymulticlass", high_card = vec, type = "multiclass", 
      remove_high_card = TRUE)
}
\references{
\url{http://healthcareai-r.readthedocs.io}

\url{https://dl.acm.org/citation.cfm?id=507538}
}
\seealso{
\code{\link{healthcareai}}
}
