% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{Title}
\usage{
prep_data(d = NULL, ..., rec_obj = NULL, convert_0_1_to_factor = TRUE,
  convert_dates = TRUE, collapse_rare_factors = TRUE,
  remove_near_zero_variance = TRUE, impute = TRUE, center = FALSE,
  scale = FALSE, dummies = FALSE, verbose = FALSE)
}
\arguments{
\item{convert_dates}{Logical or character. If TRUE (default), day-of-week,
month, and year columns are generated from date columns and date columns
are removed. If FALSE, date columns are left intact. If a character vector,
it is passed to the `features` argument of `recipes::step_date`. E.g. if
you want only month and year back: `convert_dates = c("quarter", "year")`.}

\item{verbose}{}
}
\description{
Title
}
