% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{RemoveColsWithAllSameValue}
\alias{RemoveColsWithAllSameValue}
\title{Remove columns from a data frame when those columns have the same values in
each row}
\usage{
RemoveColsWithAllSameValue(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A data frame with those columns removed
}
\description{
Remove columns from a data frame when all of their rows are the
same value
}
\examples{
df = data.frame(a=c(1,1,1),b=c('a','b','b'),c=c('a','a','a'),d = c(NA,'1',NA))
resdf = RemoveColsWithAllSameValue(df)
}
\references{
\url{https://community.healthcatalyst.com/community/data-science}
}
\seealso{
\code{\link{HCRTools}}
}

