% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-column-row-expectations.R
\name{removeColsWithDTSSuffix}
\alias{removeColsWithDTSSuffix}
\title{Remove columns with DTS suffix}
\usage{
removeColsWithDTSSuffix(df)
}
\arguments{
\item{df}{A data frame to be altered}
}
\value{
The input data frame with DTS suffix columns removed
}
\description{
Remove columns with DTS in the suffix of the column name
}
\examples{
df <- data.frame(testDTS=c(1,2,3),b=c('Y','N',NA),c=c(NA,'Y','N'))
dfResult <- removeColsWithDTSSuffix(df)
head(dfResult)
}
\references{
\url{http://healthcareai-r.readthedocs.io}
}
\seealso{
\code{\link{healthcareai}}
}
