% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-column-row-expectations.R
\name{removeColsWithOnlyNA}
\alias{removeColsWithOnlyNA}
\title{Remove columns from a data frame that are only NA}
\usage{
removeColsWithOnlyNA(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A data frame with columns of only NA's removed
}
\description{
Remove columns from a data frame when all of their rows are NA's
}
\examples{
df <- data.frame(a=c(1,1,1),
                b=c('a','b','b'),
                c=c('a','NA','a'),
                d=c(NA,NA,NA))
dfResult <- removeColsWithOnlyNA(df)
head(dfResult)

}
\references{
\url{http://healthcare.ai}
}
\seealso{
\code{\link{healthcareai}}
}
