% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{RemoveRowsWithNAInSpecCol}
\alias{RemoveRowsWithNAInSpecCol}
\title{Remove rows where specified col is NA}
\usage{
RemoveRowsWithNAInSpecCol(df, desiredCol)
}
\arguments{
\item{df}{A dataframe to be altered}

\item{desiredCol}{A column name in the df (in ticks)}
}
\value{
The input dataframe with rows removed
}
\description{
Remove rows from a dataframe where a particular col is NA
}
\examples{
df = data.frame(a=c(1,2,3),b=c('Y','N',NA),c=c(NA,'Y','N'))
resdf = RemoveRowsWithNAInSpecCol(df,'b')
}
\references{
\url{https://community.healthcatalyst.com/community/data-science}
}
\seealso{
\code{\link{HCRTools}}
}

