% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingness.R
\name{replace.missingness}
\alias{replace.missingness}
\title{Replaces missing data with NA}
\usage{
replace.missingness(d, to_replace)
}
\arguments{
\item{d}{A dataframe or matrix}

\item{to_replace}{The value that will be replaced.}
}
\value{
A dataframe with the missing data replaced.
}
\description{
This function replaces any given value in a dataframe with NA.
}
