% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingness.R
\name{replace_missingness}
\alias{replace_missingness}
\title{Replaces missing data with NA}
\usage{
replace_missingness(d, to_replace)
}
\arguments{
\item{d}{A dataframe or tibble}

\item{to_replace}{A vector of the values or string of the value that will be
replaced with NA}
}
\value{
A tibble with the missing data replaced with NA
}
\description{
This function replaces any given value in a dataframe or tibble with NA. It
can be used when missing values are filled with characters like "missing".
}
\examples{
\dontrun{
# Replace "missing" in dat
dat <- tibble(gender = c("male", "male", "female", "male", "missing"),
              age = c(64, 52, 75, "missing", 70),
              weight = c(139, 0, 193, 158, 273))

replace_missingness(dat, "missing")

# Replace both "missing" and Nan in dat
dat <- tibble(gender = c("male", NaN, "female", "male", "missing"),
              age = c(64, 52, 75, "missing", 70),
              weight = c(139, 0, 193, NaN, 273))

replace_missingness(dat, c("missing", NaN))
}

}
