% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-column-row-expectations.R
\name{returnColsWithMoreThanFiftyCategories}
\alias{returnColsWithMoreThanFiftyCategories}
\title{Return vector of columns in a data frame with greater than 50 categories}
\usage{
returnColsWithMoreThanFiftyCategories(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A vector that contains the names of the columns with greater
than 50 categories
}
\description{
Returns a vector of the names of the columns that have more than
50 categories
}
\examples{
df <- data.frame(a=c(1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,
                    1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1),
                 b=c('a','b','c','d','e','f','g','h','i','j','k','l','m','n',
                     'o','p','q','r','s','t','u','v','w','x','y','z','aa',
                     'bb','cc','dd','ee','ff','gg','hh','ii','jj','kk','ll',
                     'mm','nn','oo','pp','qq','rr','ss','tt','uu','vv','ww',
                     'xx','yy'))
colList <- returnColsWithMoreThanFiftyCategories(df)
colList
}
\references{
\url{http://healthcareai-r.readthedocs.io}
}
\seealso{
\code{\link{healthcareai}}
}
