% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{ReturnNumColsWithMoreThanFiftyFactors}
\alias{ReturnNumColsWithMoreThanFiftyFactors}
\title{Return number of columns in a dataframe with greater than 50 categories}
\usage{
ReturnNumColsWithMoreThanFiftyFactors(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
numCols An integer that is the number of columns with greater than 50 categories
}
\description{
Return number of columns in a dataframe with greater than 50 categories
}
\examples{
df = data.frame(a=c(1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,
                    1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1),
                b=c('a','b','c','d','e','f','g','h','i','j','k','l','m','n',
                    'o','p','q','r','s','t','u','v','w','x','y','z','aa','bb',
                    'cc','dd','ee','ff','gg','hh','ii','jj','kk','ll','mm','nn',
                   'oo','pp','qq','rr','ss','tt','uu','vv','ww','xx','yy'))
numCols = ReturnNumColsWithMoreThanFiftyFactors(df)
}
\seealso{
\code{\link{HCRTools}}
}

