% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantitative_interactions.R
\name{set_obs}
\alias{set_obs}
\title{set_obs makes sure that cur_observation contains all the columns necessary. It
will include any columns not provided}
\usage{
set_obs(d, cur_observation)
}
\description{
set_obs makes sure that cur_observation contains all the columns necessary. It
will include any columns not provided
}
