% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_setup.R
\name{setup_training}
\alias{setup_training}
\title{Everything in the return list gets assigned in the calling environment (tune_models and flash_models)}
\usage{
setup_training(d, outcome, model_class, models, metric, positive_class, n_folds)
}
\description{
Everything in the return list gets assigned in the calling environment (tune_models and flash_models)
}
