% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulate Counterfactual Predictions}
\usage{
simulate(models, vary = 4, hold = list(numerics = median, characters =
  Mode), numerics = 5, characters = unique)
}
\arguments{
\item{vary}{Which variables to vary? Default is 4; if it is a single integer
(n), the n most important variables are varied (see details). If it is a
vector of integers, those rankings of variables are used (e.g. \code{vary =
2:4} varies the 2nd, 3rd, and 4th most-important variables). Alternatively,
a vector of variable names to vary may be provided. Or, for the finest
level of control, a list with names being variable names and entries being
values to use; in this case \code{numerics} and
\code{characters} are ignored.}

\item{hold}{How to choose the values of variables not being varied? If this
is an integer, the values from that row in the training dataset are used.
Otherwise, it is a length-2 list with names "numerics" and "characters"
each giving functions for how to handle that data type. The default is
list(numerics = median, characters = Mode)}

\item{numerics}{For numeric variables being varied, how to select values. By
default, the minimum, 25th-, 50th-, 75th-percentile, and maximum values
from the training dataset will be used. If this is an integer, it specifies
the number of evenly spaced quantiles to use. If it is a numeric vector, it
should be in [0, 1] and specifies the quantiles to use. A function that
takes a numeric vector (the column from training data) and returns a
numeric vector (the values to be used) may also be provided}

\item{characters}{For categorical variables being varied, how to select
values? By default all unique values are used. If an integer (n) is
provided, the n-most common categories in the training dataset will be
used. A function that takes a character vector (the column from training
data) and returns a character vector (the values to be used) may also be
provided}

\item{model}{A model_list object. The data the model was trained on must have
been prepared, either by training with \code{\link{machine_learn}} or by
preparing with \code{\link{prep_data}}}
}
\value{
A tibble with values of variables used to make predictions and
  predictions. Has class \code{simulated_df} and \code{predicted_df}.
}
\description{
Simulate Counterfactual Predictions
}
\details{
If \code{vary} is an integer, the most important variables are
  determined by \code{\link{get_variable_importance}}, unless glm is the only
  model present, in which case \code{\link{interpret}} is used with a
  warning.

  When selecting the most important variables to vary over, for categorical
  variables the sum of variable importance of all the levels as dummy
  variables is used.
}
