% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-local-interpretability.R
\name{singleFactorVariableDf}
\alias{singleFactorVariableDf}
\title{Build a dataframe whose rows all agree except in one categorical column.}
\usage{
singleFactorVariableDf(baseRow, variable, factorLevels)
}
\arguments{
\item{baseRow}{a row of data (in a dataframe) to perturb}

\item{variable}{the name of the categorical column which will take on
multiple values}

\item{factorLevels}{a list of factor levels or categories which the
categorical column can take}
}
\value{
a dataframe whose rows all agree except in the \code{variable}
column which takes each level in \code{factorLevels} exactly once.
}
\description{
Build a dataframe whose rows all agree except in one categorical 
column which has a unique category in each row.
}
\references{
\url{http://healthcareai-r.readthedocs.io}
}
\seealso{
\code{\link{healthcareai}}
}
