% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-local-interpretability.R
\name{singleNumericVariableDf}
\alias{singleNumericVariableDf}
\title{Build a dataframe whose rows all agree except in one numeric column.}
\usage{
singleNumericVariableDf(baseRow, variable, interval, size = 50,
  center = NULL)
}
\arguments{
\item{baseRow}{a row of data (in a dataframe) to perturb}

\item{variable}{the name of the numeric column which will take on multiple
values}

\item{interval}{a 2-dimensional vector consisting of the endpoints of an 
interval along which to vary the value of \code{variable}}

\item{size}{the number of rows the dataframe should have}

\item{center}{a optional value to use as the center of the interval; if a 
value is specified, half of the values of the \code{variable} column will be 
less than \code{center} and the other half will be greater than \code{center}}
}
\value{
a dataframe dataframe whose rows all agree except in the
\code{variable}column whose values are evenly spread out along a range of values
}
\description{
Build a dataframe whose rows all agree except in one numeric 
column whose values are evenly spread out along a range of values.
}
\references{
\url{http://healthcareai-r.readthedocs.io}
}
\seealso{
\code{\link{healthcareai}}
}
