% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-testing-utils.R
\name{skip_if_no_MSSQL}
\alias{skip_if_no_MSSQL}
\title{Function to skip specific tests if MSSQL/specific databases not set up on 
user's machine.}
\usage{
skip_if_no_MSSQL(tableName, connString)
}
\arguments{
\item{tableName}{What table to look for if connection is good. Entered as a 
string.}

\item{connString}{Connection string to use when trying to connect to MSSQL}
}
\description{
This function is used in the testing files where R is used to 
write to MSSQL.  It first checks for a connection to MSSQL.  If no connection
to MSSQL, it will skip the test.  If there is a good connection but the
specific database is not present it will skip the test as well.
}
\references{
\url{http://healthcareai-r.readthedocs.io}
}
\seealso{
\code{\link{healthcareai}}
}
