% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-testing-utils.R
\name{skip_on_not_appveyor}
\alias{skip_on_not_appveyor}
\title{Function to skip specific tests if they are not being run on Appveyor.}
\usage{
skip_on_not_appveyor()
}
\description{
This function will skip a test if it's not being run on Appveyor. 
Used for SQL Server related tests, since we don't know if the environment will be
present. These tests are are run on Appveyor.
}
\references{
\url{http://healthcareai-r.readthedocs.io}
}
