% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_train_test.R
\name{split_train_test}
\alias{split_train_test}
\title{Split data into training and test data frames}
\usage{
split_train_test(d, outcome, percent_train = 0.8, seed, grouping_col)
}
\arguments{
\item{d}{Data frame}

\item{outcome}{Target column, unquoted. Split will be stratified across this
variable}

\item{percent_train}{Proportion of rows in d to put into training. Default is 0.8}

\item{seed}{Optional, if provided the function will return the same split
each time it is called}

\item{grouping_col}{column name that specifies grouping. Individuals in the same
group are in the same training/test set.}

\item{recipe}{A recipe object.}
}
\value{
A list of two data frames with names train and test
}
\description{
`split_train_test` splits data for validation of models. It also
offers an implentation to group all observations with the same column value
in the same train/test group. This has been created to group patients with
multiple observations in the same group.

`hcai-impute` adds various imputation methods to an existing
recipe. Currently supports mean (numeric only), new_category (categorical
only), bagged trees, or knn.
}
\details{
This function wraps `caret::createDataPartition`. If outcome is a factor
  then the test/training porportions are stratified. Otherwise they are randomly
  selected.

  If the grouping_col is given, then the groups are divided into the test/
  training porportions.
}
\examples{
split_train_test(mtcars, am, .9)


# splits to training and tests while grouping on the owner of cars
mtcars_owners <-
  mtcars \%>\%
  mutate(owner = as.factor(rep(letters[1:16], each = 2)))

split_train_test(mtcars_owners, am, grouping_col = owner)

# Grouping patients
split_train_test(d, outcome, grouping_col = patient_id)

}
