% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-logHelpers.R
\name{startProdLogs}
\alias{startProdLogs}
\title{Sets console logging to a file in the working directory.}
\usage{
startProdLogs()
}
\description{
Logs all console messages and output to a timestamped file called, "deployConsoleLog..."
in the working directory. This is meant to be used in production to monitor console output on ETL servers.
}
\examples{
# Object Attribute Base should contain a script like this:
# Move to working directory 
# setwd("Prod Server/Machine Learning/MyProject")

# Start console logging
startProdLogs()

# Deploy -------------------------------------------------
# Test deployment. Comment/delete when working
catalyst_test_deploy_in_prod()

# Deploy model
# source("myDeployScript.R")

# Stop console logging
stopProdLogs()
}
\seealso{
\code{\link{stopProdLogs}}
}
