% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_hcai_missing.R
\name{step_hcai_missing}
\alias{step_hcai_missing}
\title{Clean NA values from categorical/nominal variables}
\usage{
step_hcai_missing(recipe, ..., role = NA, trained = FALSE,
  na_percentage = NULL)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of 
operations for this recipe.}

\item{...}{One or more selector functions to choose which variables are 
affected by the step. See [selections()] for more details.}

\item{role}{Not used by this step since no new variables are created.}

\item{trained}{A logical to indicate if the number of NA values have been
counted in preprocessing.}

\item{na_percentage}{A named numeric vector of NA percentages. This
is `NULL` until computed by [prep.recipe()].}
}
\value{
An updated version of `recipe` with the new step
 added to the sequence of existing steps (if any). For the
 `tidy` method, a tibble with columns `terms` (the
 selectors or variables selected) and `value` (the
 NA counts).
}
\description{
`step_hcai_missing` creates a specification of a recipe that 
 will replace NA values `hcai_missing`.
}
\details{
NA values must g `step_scale` estimates
 the variable standard deviations from the data used in the
 `training` argument of `prep.recipe`.
 `bake.recipe` then applies the scaling to new data sets
 using these standard deviations.
}
\examples{
df <- data.frame(date=c('2009-01-01','2010-01-01','2009-03-08','2009-01-19'),
                a=c(1,2,3,4))
dfResult <- orderByDate(df,'date', descending=FALSE)
head(dfResult)
}
