% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_locfimpute.R
\name{step_locfimpute}
\alias{step_locfimpute}
\alias{tidy.step_locfimpute}
\title{Last Observation Carried Forward}
\usage{
step_locfimpute(recipe, ..., role = NA, trained = FALSE, skip = FALSE)

\method{tidy}{step_locfimpute}(x, ...)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of
operations for this recipe.}

\item{...}{One or more selector functions to choose which variables will be
imputed. See [selections()] for more details. For the `tidy` method, these
are not currently used.}

\item{role}{Not used by this step since no new variables are
created.}

\item{trained}{A logical to indicate if the number of NA values have been
counted in preprocessing.}

\item{skip}{A logical. Should the step be skipped when the recipe is baked?}

\item{x}{A `step_locfimpute` object.}
}
\value{
For `step_locfimpute`, an updated version of recipe with the new step
  added to the sequence of existing steps (if any). For the `tidy` method, a
  tibble with columns `terms` (the selectors or variables selected), `value` ### GOING TO NEED TO CHANGE THIS AFTER I FIGURE OUT TIDY
  (the feature names), and `ordinal` (a logical).
}
\description{
`step_locfimpute` creates a *specification* of a recipe step that
 will convert date data into factor or numeric variable(s). This step will
 guess the date format of columns with the "_DTS" suffix, and then create
 either `categories` or `continuous` columns. Various portions of this step
 are copied from `recipes::step_date`.
}
\details{
`step_locfimpute` imputes by using the last observation and carrying
  it forward. If the first observation is missing, it is imputed with the
  first observation.
}
\examples{
library(recipes)

prepped <-
  recipe(formula = "~.", pima_diabetes) \%>\%
  step_locfimpute(weight_class, insulin, skinfold, diastolic_bp) \%>\%
  prep()

bake(prepped, newdata = pima_diabetes)
}
