% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{stop_prod_logs}
\alias{stop_prod_logs}
\title{Stops all console logging.}
\usage{
stop_prod_logs()
}
\description{
Stops all console logging. This is meant to be used in
production to monitor console output on ETL servers.
}
\seealso{
\code{\link{start_prod_logs}}
}
