% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-logHelpers.R
\name{stopProdLogs}
\alias{stopProdLogs}
\title{Stops all console logging.}
\usage{
stopProdLogs()
}
\description{
Stops all console logging. This is meant to be used in production to monitor console output on ETL servers.
}
\seealso{
\code{\link{startProdLogs}}
}
