% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupervisedModel.R
\docType{class}
\name{SupervisedModel}
\alias{SupervisedModel}
\title{Compare predictive models, created on your data}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SupervisedModel
}
\arguments{
\item{object}{of SuperviseModelParameters class for $new() constructor}
}
\description{
This step allows one to create test models on your data
and helps determine which performs best.
}
\references{
\url{http://healthcareml.org/}
}
\seealso{
\code{\link{HCRTools}}
}
\keyword{datasets}

