% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervised-model-deployment.R
\docType{class}
\name{SupervisedModelDeployment}
\alias{SupervisedModelDeployment}
\title{Deploy predictive models, created on your data}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SupervisedModelDeployment(object)
}
\arguments{
\item{object}{of SupervisedModelDeploymentParams class for $new() constructor}
}
\description{
This step allows one to create deploy models on your data
and helps determine which performs best.
}
\references{
\url{http://healthcareai-r.readthedocs.io}
}
\seealso{
\code{\link{healthcareai}}
}
\keyword{datasets}
