% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervised-model-deployment-params.R
\docType{class}
\name{SupervisedModelDeploymentParams}
\alias{SupervisedModelDeploymentParams}
\title{SupervisedModelDeploymentParams class to set up parameters required to build
SupervisedModelDeployment class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SupervisedModelDeploymentParams
}
\description{
This step allows one to create deploy models on your data
and helps determine which performs best.
}
\references{
\url{http://healthcareai-r.readthedocs.io}
}
\seealso{
\code{\link{healthcareai}}
}
\keyword{datasets}
