% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervised-model-development.R
\docType{class}
\name{SupervisedModelDevelopment}
\alias{SupervisedModelDevelopment}
\title{Compare predictive models, created on your data}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SupervisedModelDevelopment(object)
}
\arguments{
\item{object}{of SuperviseModelParameters class for $new() constructor}
}
\description{
This step allows one to create test models on your data
and helps determine which performs best.
}
\references{
\url{http://healthcareai-r.readthedocs.io}
}
\seealso{
\code{\link{healthcareai}}
}
\keyword{datasets}
