% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervised-model-development-params.R
\docType{class}
\name{SupervisedModelDevelopmentParams}
\alias{SupervisedModelDevelopmentParams}
\title{SupervisedModelDevelopmentParams class to set up parameters required to build SupervisedModel classes}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SupervisedModelDevelopmentParams
}
\description{
This step allows one to create test models on your data
and helps determine which performs best.
}
\references{
\url{http://healthcareai-r.readthedocs.io}
}
\seealso{
\code{\link{healthcareai}}
}
\keyword{datasets}
