% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupervisedModelParameters.R
\docType{class}
\name{SupervisedModelParameters}
\alias{SupervisedModelParameters}
\title{SupervisedModelParameters class to set up parameters required to build SupervisedModel classes}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SupervisedModelParameters
}
\description{
This step allows one to create test models on your data
and helps determine which performs best.
}
\references{
\url{http://products.healthcatalyst.com/Predictive}
}
\seealso{
\code{\link{HCRTools}}
}
\keyword{datasets}

