% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised-model.R
\docType{class}
\name{UnsupervisedModel}
\alias{UnsupervisedModel}
\title{Compare predictive models, created on your data}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
UnsupervisedModel(object)
}
\arguments{
\item{object}{of UnsupervisedModelParams class for $new() constructor}
}
\description{
This step allows one to create test models on your data
and helps determine which performs best.
}
\references{
\url{http://healthcare.ai}
}
\seealso{
\code{\link{healthcareai}}
}
\keyword{datasets}
