% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised-model-params.R
\docType{class}
\name{UnsupervisedModelParams}
\alias{UnsupervisedModelParams}
\title{UnsupervisedModelParams class to set up parameters required to build UnsupervisedModel classes}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
UnsupervisedModelParams
}
\description{
This step allows one to create test models on your data
and helps determine which performs best.
}
\references{
\url{http://healthcare.ai}
}
\seealso{
\code{\link{healthcareai}}

\code{\link{KmeansClustering}}
}
\keyword{datasets}
