% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connections.R
\name{writeData}
\alias{writeData}
\title{Depreciated.}
\usage{
writeData(MSSQLConnectionString = NULL, df, SQLiteFileName = NULL,
  tableName, addSAMUtilityColumns = FALSE, connectionString = NULL)
}
\description{
Removed in v2.0.0
}
