% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost-deployment.R
\docType{class}
\name{XGBoostDeployment}
\alias{XGBoostDeployment}
\title{Deploy a production-ready predictive RandomForest model}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
XGBoostDeployment(type, df, grainCol, testWindowCol, 
predictedCol, impute, debug)
}
\arguments{
\item{type}{The type of model (either 'regression' or 'classification')}

\item{df}{Dataframe whose columns are used for calc.}

\item{grainCol}{The dataframe's column that has IDs pertaining to the grain}

\item{testWindowCol}{Y or N. This column dictates the split between model 
training and test sets. Those rows with N in this column indicate the 
training set while those that have Y indicate the test set}

\item{predictedCol}{Column that you want to predict. If you're doing
classification then this should be Y/N.}

\item{impute}{For training df, set all-column imputation to F or T.
This uses mean replacement for numeric columns
and most frequent for factorized columns.
F leads to removal of rows containing NULLs.}

\item{debug}{Provides the user extended output to the console, in order
to monitor the calculations throughout. Use T or F.}
}
\description{
This step allows one to
\itemize{
\item Load a saved model from \code{\link{RandomForestDevelopment}}
\item Run the model against test data to generate predictions
\item Push these predictions to SQL Server
}
}
\examples{

}
\seealso{
\code{\link{healthcareai}}
}
\keyword{datasets}
