% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_list_constructors.R
\name{as.model_list}
\alias{as.model_list}
\title{Make models into model_list object}
\usage{
as.model_list(..., listed_models = NULL, target = ".outcome", model_class,
  tuned = TRUE, recipe = NULL, positive_class = NULL)
}
\arguments{
\item{...}{\code{caret}-trained models to put into a model list}

\item{listed_models}{Use this if your models are already in a list}

\item{target}{Quoted name of response variable}

\item{model_class}{"classification" or "regression". Will be determined if
not provided}

\item{tuned}{Logical; if FALSE, will have super-class untuned_models}

\item{recipe}{recipe object from prep+_data, or NULL if the data didn't go
through prep_data}

\item{positive_class}{If classification, the positive outcome class,
otherwise NULL}
}
\value{
A model_list with child class type_list
}
\description{
Make models into model_list object
}
