% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pip.R
\name{build_one_level_df}
\alias{build_one_level_df}
\title{Replace all value in the column of a dataframe with a given value.}
\usage{
build_one_level_df(dataframe, variable, level, one_variable_direction)
}
\arguments{
\item{dataframe}{a dataframe}

\item{variable}{The name of the column whose values we wish to
change}

\item{level}{The modified value to use in the \code{variable} column}
}
\value{
A dataframe
}
\description{
Takes a dataframe and replaces all the values in the
\code{variable} column with the value \code{level}. Columns
tracking which variable and level were used are also added to the dataframe.
}
\keyword{internal}
