% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-correlations.R
\name{calculateAllCorrelations}
\alias{calculateAllCorrelations}
\title{Correlation analysis on an input table over all numeric columns}
\usage{
calculateAllCorrelations(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A data frame with column names and corresponding correlations
with the target column
}
\description{
Calculate correlations between every numeric column in a table
}
\examples{
df <- data.frame(a=c(1,2,3,4,5,6),
b=c(6,5,4,3,2,1),
c=c(3,4,2,1,3,5),
d=c('M','F','F','F','M','F')) #<- is ignored

dfResult <- calculateAllCorrelations(df)
dfResult
}
\references{
\url{http://healthcareai-r.readthedocs.io}
}
\seealso{
\code{\link{healthcareai}}
}
