% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{calulcateAlternatePredictions}
\alias{calulcateAlternatePredictions}
\title{Recalculate predicted value based on alternate scenarios}
\usage{
calulcateAlternatePredictions(df, modelObj, type, outVectorAppend)
}
\arguments{
\item{df}{Data frame from which we'll calculate alternate predictions}

\item{modelObj}{Object representing the model that is used for predictions}

\item{type}{String representing which type of model is used}

\item{outVectorAppend}{Optional list of values that we'll append predictions
to. If not used, then a new vector is created.}
}
\description{
After getting alternate features via calculateSDChanges
recalculate predicted values for each row in df.
}
\references{
\url{http://healthcare.ai/}
}
\seealso{
\code{\link{HCRTools}}
}

