% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-date-time.R
\name{convertDateTimeColToDummies}
\alias{convertDateTimeColToDummies}
\title{Converts datetime columns into dummy columns}
\usage{
convertDateTimeColToDummies(df, dateTimeCol, depth = "h",
  returnDtCol = FALSE)
}
\arguments{
\item{df}{A data frame. Indicates the datetime column.}

\item{dateTimeCol}{A string. Column name in df that will be converted
into several columns.}

\item{depth}{A string. Specifies the depth with which to expand extra columns
(starting with a year column). 'd' expands to day, 'h' expands to hour
(default), m' expands to minute, and 's' expands to second.}

\item{returnDtCol}{A boolean. Return the original dateTimeCol with
the modified data frame?}
}
\value{
Warning message stating the function has been renamed to 
splitOutDateTimeCols.
}
\description{
THIS FUNCTION HAS BEEN RENAMED TO splitOutDateTimeCols
}
\references{
\url{http://healthcareai-r.readthedocs.io}
}
\seealso{
\code{\link{healthcareai}}
}
