% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-feature-availability-profiler.R
\name{countPercentEmpty}
\alias{countPercentEmpty}
\title{DEPRECATED. Calculates percentage of each column in df that is NULL (NA)}
\usage{
countPercentEmpty(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A vector that contains the percentage of NULL in each column
}
\description{
Returns a vector with percentage of each column that is NULL
in the original data frame
}
\examples{
df = data.frame(a=c(1,2,NA,NA,3),
                b=c(NA,NA,NA,NA,NA),
                c=c(NA,NA,'F','M',NA))
colList = countPercentEmpty(df)
colList
}
\references{
\url{http://healthcareai-r.readthedocs.io}
}
\seealso{
\code{\link{healthcareai}}
}
