% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-variation.R
\name{createVarianceTallTable}
\alias{createVarianceTallTable}
\title{Transform a dataframe to be three columns and tall instead of wide}
\usage{
createVarianceTallTable(df, categoricalCols, measure)
}
\arguments{
\item{df}{A dataframe}

\item{categoricalCols}{Vector of categorical column(s)}

\item{measureColumn}{A string. The measure of interest.}
}
\value{
A dataframe of four columns
}
\description{
When dealing with a table that could be unexpectedly wide,
it helps to instead fix its width and let it get tall (which makes it easy
to insert into a pre-existing table). First use findVariation function.
}
\examples{
df <- data.frame(LactateOrderProvSpecialtyDSC = c("Pulmonary Disease",
                                                  "Family Medicine"),
                 LactateOrderProvNM = c("Hector Salamanca",
                                        "Gus Fring"),
                 LOSCOV = c(43.4,35.1),
                 LOSVolume = c(2,3),
                 LOSImpact = c(86.8,105.3))

head(df)

categoricalCols <- c("LactateOrderProvSpecialtyDSC","LactateOrderProvNM")

dfRes <- createVarianceTallTable(df = df, 
                                 categoricalCols = categoricalCols, 
                                 measure = "LOS")
head(dfRes)
}
\references{
\url{http://healthcare.ai}
}
\seealso{
\code{\link{healthcareai}} 
\code{\link{findVariation}}
}
