% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connections.R
\name{db_write}
\alias{db_write}
\title{Write to a SQL Server database table}
\usage{
db_write(d, con, schema = "dbo", table_name, overwrite = FALSE)
}
\arguments{
\item{d}{A data frame. This data will be written to the specified table. Data
must match the destination table in number of columns, names of columns, and
data types.}

\item{con}{An odbc database connection. Can be made using
\code{build_connection_string}. This connection must specify a database.}

\item{schema}{String representing the table schema. database.schema.table_name
must already exist. Schema defaults to "dbo".}

\item{table_name}{String representing the table name. database.schema.table_name
must already exist.}

\item{overwrite}{Logical. If FALSE (default), data will be appended to table.
If TRUE, table will be overwritten.}
}
\value{
Silently returns confirmation that data was written correctly.
}
\description{
Use a database connection to write to an existing SQL Server
table. Data types and columns must match exactly.
}
\details{
This function requires that: \enumerate{
\item{A database is specified in the database connection}
\item{Data matches destination table in number, name, and types of columns}
\item{The destination table, made from the connection's database, specified
schema, and specified table name, must exist.}
}
This function will not create a new table.
}
\examples{
\dontrun{
my_con <- build_connection_string(server = "HPHI-EDWDEV",
                                  database = "SAM")
con <- DBI::dbConnect(odbc::odbc(), .connection_string = my_con)
d <- tibble(id = 123, name = John Smith, sex = "Male")
d <- db_write(d,
              con,
              schema = "patient",
              table_name = "demographics")
}

}
\seealso{
\code{\link{build_connection_string}}

\code{\link{db_read}}
}
