% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-kmeans.R
\name{distancePointLine}
\alias{distancePointLine}
\title{Compute the distance of a point from a line}
\usage{
distancePointLine(x, y, slope, intercept)
}
\arguments{
\item{x}{A numeric vector, x axis of the points}

\item{y}{A numeric vector, y axis of the points}

\item{slope}{A number, the slope of the line}

\item{intercept}{A number, the intercept of the line}
}
\value{
A numeric vector, the length of the pointS from the line.
}
\description{
compute the distance of a point from a line given the 
x, y axes of the point and the slope and intercept of the line.
}
\examples{
distancePointLine(c(2,0),c(3,2),-0.5,2.5)
}
\references{
\url{http://healthcare.ai}

This helper funtion is originally found here 
\url{http://paulbourke.net/geometry/pointlineplane/pointline.r}

\url{https://github.com/bryanhanson/ChemoSpecMarkeR/blob/master/R/findElbow.R}
}
\seealso{
\code{\link{healthcareai}}
}
