% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{FindTrendsAboveThreshold}
\alias{FindTrendsAboveThreshold}
\title{Find any columns that have a trend above a particular threshold}
\usage{
FindTrendsAboveThreshold(df, datecol, threshold = 0.5, nelson = TRUE)
}
\arguments{
\item{df}{A dataframe}
}
\value{
A vector of column names
}
\description{
Find numeric columns in dataframe that have an absolute slope greater than
that specified via threshold argument.
}
\examples{
x <- seq(as.Date("2012-01-01"), as.Date("2012-01-06"), by = "days")
y1 <- c(1,3,6,8,13,14)          # big positive
y2 <- c(1,1.2,1.2,1.2,1.3,1.3)  # small positive
y3 <- c(0,-2,-2,-4,-5,-7)       # big negative
y4 <- c(0,-.5,-.5,-.5,-.5,-.6)  # small negative
y5 <- c(0,1,1,2,5,5)            # factor col

data <- data.frame(x,y1,y2,y3,y4,y5)
data$y5 <- as.factor(data$y5)

col_list <- FindTrendsAboveThreshold(d = data,
                                     datecol = 'x',
                                     threshold = 0.5)
}
\seealso{
\code{\link{HCRTools}}
}

