% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-variation.R
\name{findVariation}
\alias{findVariation}
\title{Find high variation}
\usage{
findVariation(df, categoricalCols, measureCol, dateCol = NULL,
  threshold = NULL)
}
\arguments{
\item{df}{A dataframe}

\item{categoricalCols}{Vector of categorical columns}

\item{measureCol}{A string of the measure column of interest}

\item{dateCol}{Optional. A date(time) column to group by (done by month)}

\item{threshold}{A number, representing the minimum coefficient of variance
(COV) to flag and surface}
}
\value{
A dataframe of three columns
}
\description{
Search across subgroups and surface those that have variation
above a particular threshold
}
\examples{
df <- data.frame(Dept = c('A','A','A','B','B','B','B'),
                 Gender = c('F','M','M','M','M','F','F'),
                 LOS = c(3.2,NA,5,1.3,2.4,4,9))

categoricalCols <- c("Dept","Gender")

df <- findVariation(df = df, 
                    categoricalCols = categoricalCols,
                    measureCol = "LOS")

# Only subcategories w/ more than one row (with non-NA measure) are eligible
df
}
\references{
\url{http://healthcare.ai}
}
\seealso{
\code{\link{healthcareai}} \code{\link{calculateCOV}} 
\code{\link{createVarianceTallTable}}
}
