% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{generateAUC}
\alias{generateAUC}
\title{Generate ROC curve for a dataset.}
\usage{
generateAUC(predictions, labels, aucType, plotFlg)
}
\arguments{
\item{predictions}{A vector of predictions from a machine learning model.}

\item{labels}{A vector of the true labels. Must be the same length as 
predictions.}

\item{aucType}{Can be "SS" or "PR". Defaults to SS.}

\item{plotFlg}{Binary value controlling plots Defaults to FALSE (no).}
}
\value{
auc Integral AUC of ROC plot.
}
\description{
Generates ROC curve and AUC for Sensitivity/Specificity or 
Precision/Recall.
}
\examples{

# generate data
df <- data.frame(a = rep( seq(0,1,by=0.1), times=5),
                 b = rep(0,times=55))
# add positive labels
df[c(seq(11,55,by=11)),2] <- 1
df[c(seq(10,55,by=11)),2] <- 1
df[c(seq(9,55,by=22)),2] <- 1
df[c(4, 18, 30, 39, 52 ),2 ] <- 1
# prepare vectors
pred <- df[,1]
labels <- df[,2]

# generate the AUC
auc = generateAUC(pred,labels,'SS','TRUE')

}
\references{
\url{http://healthcare.ai}
}
\seealso{
\code{\link{healthcareai}}
}

