% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tune_grid.R
\name{get_hyperparameter_defaults}
\alias{get_hyperparameter_defaults}
\title{Get default hyperparameter lists that can be passed to flash_models}
\usage{
get_hyperparameter_defaults(models, n, k, model_class)
}
\arguments{
\item{n}{Number observations}

\item{k}{Number features}

\item{model_class}{"classification" or "regression"}

\item{model}{which algorithms?}
}
\value{
List of named lists that can be passed to \code{\link{flash_models}}
}
\description{
Get default hyperparameter lists that can be passed to flash_models
}
\details{
Get default hyperparameters to use, e.g. in flash_models. list
  (algorithms) of lists (hyperparameters).

  k-NN defaults are from the kknn package: kmax = 7, distance = 2
  (Minkowski's exponent, i.e. Euclidean distance), kernal = "optimal"

  Random forest defaults are from Intro to Statistical Learning and caret:
  mtry = sqrt(k), splitrule = "extratrees", min.node.size = 1 for
  classification, 5 for regression
}
