% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_hyperparameter_defaults}
\alias{get_hyperparameter_defaults}
\title{get_hyperparameter_defaults}
\usage{
get_hyperparameter_defaults(models = get_supported_models())
}
\arguments{
\item{models}{Character vector of algorithms for which hyperparameters are
desired.}
}
\value{
List of named lists that can be passed to \code{\link{flash_models}}
}
\description{
get_hyperparameter_defaults
}
\details{
Get default hyperparameters to use, e.g. in flash_models. list
  (algorithms) of lists (hyperparameters).
}
