% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-model-performance.R
\name{getCutOffList}
\alias{getCutOffList}
\title{Function to return ideal cutoff and TPR/FPR or precision/recall.}
\usage{
getCutOffList(perf, aucType = "SS", allCutoffsFlg = FALSE)
}
\arguments{
\item{perf}{An ROCR performance class. (Usually made by generateAUC)}

\item{aucType}{A string. Indicates AUC_ROC or AU_PR and can be "SS" or "PR". 
Defaults to SS.}

\item{allCutoffsFlg}{Binary value controlling list of all thresholds.}
}
\value{
Array of ideal cutoff and associated TPR/FPR or pre/rec.
}
\description{
Calculates ideal cutoff by proximity to corner of the ROC curve.
Usually called from \code{\link{generateAUC}}
}
\references{
\url{http://healthcareai-r.readthedocs.io}
}
\seealso{
\code{\link{healthcareai}}
}
