% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{GroupedLOCF}
\alias{GroupedLOCF}
\title{Last observation carried forward}
\usage{
GroupedLOCF(dt, id)
}
\arguments{
\item{dt}{data.table sorted by an ID column and a time or sequence number
column.}

\item{id}{A column name (in ticks) in dt to group rows by.}
}
\value{
A data.table where the last non-NA values are carried forward (overwriting
NAs) until the group ID changes.
}
\description{
Carries the last observed value forward for all columns in a data.table
grouped by an id.
}
\examples{
library(data.table)
dt = data.table(PersonID=c(1,1,2,2,3,3,3),
                wt=c(.5,NA,NA,NA,.3,.7,NA),
                ht=c(NA,1,3,NA,4,NA,NA),
                date=c('01/01/2015','01/15/2015','01/01/2015','01/15/2015',
                       '01/01/2015','01/15/2015','01/30/2015'))

head(dt,n=7)

df.result = GroupedLOCF(dt, 'PersonID')

head(df.result,n=7)
}
\seealso{
\code{\link{HCRTools}}
}

