% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCRTools.R
\docType{package}
\name{HCRTools}
\alias{HCRTools}
\alias{HCRTools-package}
\title{HCRTools: a streamlined way to develop and deploy models}
\description{
HCRTools provide a clean interface that lets one create and compare multiple
models on your data, and then deploy the model that is most accurate.
}
\details{
This is done in a two-step process:

\itemize{
\item Use \code{\link{DevelopSupervisedModel}} to test and compare models
based on your data.
\item Once you've determined which model is best, use
\code{\link{DeploySupervisedModel}} to create a final model, automatically
save it, predict against test data, and push predicted values into SQL
Server.
}
}
\seealso{
\code{\link{DevelopSupervisedModel}}

\code{\link{DeploySupervisedModel}}

\code{\link{GroupedLOCF}}

\code{\link{SelectData}}

\code{\link{IsBinary}}

\code{\link{ImputeColumn}}

\code{\link{RemoveRowsWithNAInSpecCol}}
}

