% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCRTools.R
\docType{package}
\name{HCRTools}
\alias{HCRTools}
\alias{HCRTools-package}
\title{HCRTools: a streamlined way to develop and deploy models}
\description{
HCRTools provide a clean interface that lets one create and compare multiple
models on your data, and then deploy the model that is most accurate.
}
\details{
This is done in a two-step process:

\itemize{
\item Use \code{\link{Lasso}} or \code{\link{RandomForest}} to test and
compare models based on your data.
\item Once you've determined which model is best, use
\code{\link{DeployLasso}} or \code{\link{DeployRandomForest}} to create a
final model, automatically save it, predict against test data, and push
predicted values into SQL Server.
}
}
\references{
\url{http://healthcareml.org/}
}
\seealso{
\code{\link{Lasso}}

\code{\link{RandomForest}}

\code{\link{RiskAdjustedComparisons}}

\code{\link{ImputeColumn}}

\code{\link{GroupedLOCF}}

\code{\link{SelectData}}

\code{\link{WriteData}}

\code{\link{OrderByDate}}

\code{\link{IsBinary}}

\code{\link{RemoveRowsWithNAInSpecCol}}

\code{\link{CountPercentEmpty}}

\code{\link{RemoveColsWithAllSameValue}}

\code{\link{ReturnColsWithMoreThanFiftyCategories}}

\code{\link{FindTrends}}

\code{\link{ConvertDateTimeColToDummies}}

\code{\link{CountDaysSinceFirstDate}}

\code{\link{CalculateTargetedCorrelations}}

\code{\link{CalculateAllCorrelations}}
}

