% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthcareai.R
\docType{package}
\name{healthcareai}
\alias{healthcareai}
\alias{healthcareai-package}
\title{healthcareai: a streamlined way to develop and deploy models}
\description{
healthcareai provides a clean interface that lets one create and compare multiple
models on your data, and then deploy the model that is most accurate. healthcareai
also includes functions for data exploration, data cleaning, and model evaluation.
}
\details{
This is done in a four-step process:

\enumerate{
\item{\strong{Load and profile data.}}{\cr Use \code{\link{selectData}} to pull data directly 
from the SQL database. Then, \code{\link{featureAvailabilityProfiler}} can help
determine how many null values are in a column and how they are populated over time.}

\item{Step 2:}{Build a machine learning model using \code{\link{LassoDevelopment}} or
\code{\link{RandomForestDevelopment}} and test different combinations of 
features. Determine the best model using: 
\itemize{
\item Area under the ROC curve or area under the Performance-Recall 
curve for classification problems (yes or no response). 
\item Mean squared error for regression problems (continuous response).
}}

\item{Step 3:}{Once you've determined which model is best, use
\code{\link{LassoDeployment}} or \code{\link{RandomForestDeployment}} to
create a final model, automatically save it, predict against test data, and
push predicted values into SQL Server.}
}
}
\references{
\url{http://healthcare.ai}
}
\seealso{
\code{\link{LinearMixedModelDevelopment}}

\code{\link{LinearMixedModelDeployment}}

\code{\link{RiskAdjustedComparisons}}

\code{\link{imputeColumn}}

\code{\link{groupedLOCF}}

\code{\link{selectData}}

\code{\link{writeData}}

\code{\link{orderByDate}}

\code{\link{isBinary}}

\code{\link{removeRowsWithNAInSpecCol}}

\code{\link{countPercentEmpty}}

\code{\link{removeColsWithAllSameValue}}

\code{\link{returnColsWithMoreThanFiftyCategories}}

\code{\link{findTrends}}

\code{\link{convertDateTimeColToDummies}}

\code{\link{countDaysSinceFirstDate}}

\code{\link{calculateTargetedCorrelations}}

\code{\link{calculateAllCorrelations}}
}

