% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthcareai.R
\docType{package}
\name{healthcareai}
\alias{healthcareai}
\alias{healthcareai-package}
\title{healthcareai: a streamlined way to develop and deploy models}
\description{
healthcareai provides a clean interface that lets one create and compare multiple
models on your data, and then deploy the model that is most accurate.
}
\details{
This is done in a two-step process:

\itemize{
\item Use \code{\link{LassoDevelopment}} or
\code{\link{RandomForestDevelopment}} to test and
compare models based on your data.
\item Once you've determined which model is best, use
\code{\link{LassoDeployment}} or \code{\link{RandomForestDeployment}} to
create a final model, automatically save it, predict against test data, and
push predicted values into SQL Server.
}
}
\references{
\url{http://healthcare.ai}
}
\seealso{
\code{\link{LinearMixedModelDevelopment}}

\code{\link{LinearMixedModelDeployment}}

\code{\link{RiskAdjustedComparisons}}

\code{\link{imputeColumn}}

\code{\link{groupedLOCF}}

\code{\link{selectData}}

\code{\link{writeData}}

\code{\link{orderByDate}}

\code{\link{isBinary}}

\code{\link{removeRowsWithNAInSpecCol}}

\code{\link{countPercentEmpty}}

\code{\link{removeColsWithAllSameValue}}

\code{\link{returnColsWithMoreThanFiftyCategories}}

\code{\link{findTrends}}

\code{\link{convertDateTimeColToDummies}}

\code{\link{countDaysSinceFirstDate}}

\code{\link{calculateTargetedCorrelations}}

\code{\link{calculateAllCorrelations}}
}

