% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthcareai.R
\docType{package}
\name{healthcareai}
\alias{healthcareai}
\title{Machine Learning Made Easy}
\description{
healthcare.ai makes it as easy as possible to pull data from a database, get
it ready for machine learning, optimize multiple models, and deploy
predictions.
}
\details{
The package website -- \url{https://docs.healthcare.ai/} -- contains vignettes
that demonstrate how to use the package, as well as documentation of all the
important functions.
}
