% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-impute.R
\name{imputeColumn}
\alias{imputeColumn}
\title{Depreciated in favor of \code{imputeDF}}
\usage{
imputeColumn(v)
}
\arguments{
\item{v}{A vector, or column of values with NAs.}
}
\value{
A vector, or column of values now with no NAs
}
\description{
This class performs imputation on a vector. For numeric vectors
the vector-mean is used; for factor columns, the most frequent value is used.
}
