% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-column-row-expectations.R
\name{isBinary}
\alias{isBinary}
\title{Check if a vector has only two unique values.}
\usage{
isBinary(v)
}
\arguments{
\item{v}{A vector, or column of values}
}
\value{
A boolean
}
\description{
Check if a vector is binary (not counting NA's)
}
\examples{
isBinary(c(1,2,NA))
isBinary(c(1,2,3))
}
\references{
\url{http://healthcareai-r.readthedocs.io}
}
\seealso{
\code{\link{healthcareai}}
}
