% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedModel.R
\docType{class}
\name{LinearMixedModel}
\alias{LinearMixedModel}
\title{Compare predictive models, created on your data}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LinearMixedModel
}
\description{
This step allows one to create test models on your data
and helps determine which performs best.
}
\keyword{datasets}

