% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-kmeans.R
\name{lineMagnitude}
\alias{lineMagnitude}
\title{Compute the distance between two points}
\usage{
lineMagnitude(x1, y1, x2, y2)
}
\arguments{
\item{x1}{A numeric vector, x axis of the points}

\item{y1}{A numeric vector, y axis of the points}

\item{x2}{A numeric vector, x axis of the points}

\item{y2}{A numeric vector, y axis of the points}
}
\value{
A numeric vector, the length between two points.
}
\description{
compute the distance between two points given the x, y axes of the
points.
}
\examples{
lineMagnitude(c(1,0),c(2,0),5,3)
}
\references{
\url{http://healthcare.ai}
}
\seealso{
\code{\link{healthcareai}}
}
