% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-local-interpretability.R
\name{localLinearApproximation}
\alias{localLinearApproximation}
\title{Build a linear model approximating another model}
\usage{
localLinearApproximation(baseRow, predictFunction, localDf)
}
\arguments{
\item{baseRow}{A row in a dataframe containing a data point at which to make 
the linear approximation.}

\item{predictFunction}{A function with which to make predictions for the
model to approximate}

\item{localDf}{A dataframe of data near baseRow on which to approximate the 
model}
}
\value{
The linear model approximating the original model
}
\description{
Build a linear model approximating another model at a point
}
\references{
\url{http://healthcare.ai}
}
\seealso{
\code{\link{healthcareai}}
}
