% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{OrderByDate}
\alias{OrderByDate}
\title{Return a dataframe ordered by date}
\usage{
OrderByDate(df, datecol, descending = FALSE)
}
\arguments{
\item{df}{A dataframe}

\item{descending}{Boolean for whether the output should be in descending order}

\item{dateCol}{Name of column in dataframe that contains dates}
}
\value{
df A dataframe ordered by date column
}
\description{
Returns a dataframe that's ordered by its date column
}
\examples{
library(HCRTools)
df = data.frame(date=c('2009-01-01','2010-01-01','2009-03-08','2009-01-19',
                       '2010-11-11'),
                a=c(1,2,3,4,5))
resdf = OrderByDate(df,'date',descending=FALSE)
}
\references{
\url{https://community.healthcatalyst.com/community/data-science}
}
\seealso{
\code{\link{HCRTools}}
}

