% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pip.R
\name{pip}
\alias{pip}
\title{Patient Impact Predictor}
\usage{
pip(model, d, new_values, n = 3, repeated_factors = FALSE,
  smaller_better = TRUE, expected_signs, excluded_transitions, id)
}
\arguments{
\item{model}{A model_list object}

\item{d}{A data frame on which \code{model} can make predictions}

\item{new_values}{A list of alternative values of variables of interest. The
names of the list must be variables in \code{d} and the entries are the
alternative value(s) to try.}

\item{n}{Integer, default = 3. The number of alternatives to return for each
patient.}

\item{repeated_factors}{Logical, default = FALSE. Do you want multiple
modifications of the same variable for the same patient? E.g. Reducing BMI
to 15 and to 20.}

\item{smaller_better}{Logical, default = TRUE. Are lesser values of the
outcome variable in \code{model} preferable?}

\item{expected_signs}{TODO}

\item{excluded_transitions}{TODO}

\item{id}{An identifying column in \code{d} to return in the output data
frame. If this isn't provided, an ID column from \code{model}'s data prep
will be used if available.}
}
\description{
Patient Impact Predictor
}
\examples{
m <- machine_learn(pima_diabetes, patient_id, outcome = diabetes,
                   tune = FALSE, models = "xgb")
pip(model = m, d = pima_diabetes[1:3, ],
    new_values = list(weight_class = c("underweight", "normal", "overweight"),
                      plasma_glucose = c(75, 100)))
}
