% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_list_generics.R
\name{plot.model_list}
\alias{plot.model_list}
\title{Plot performance of models}
\usage{
\method{plot}{model_list}(mlist, print = TRUE)
}
\arguments{
\item{mlist}{modellist object as returned by \code{\link{tune}}}

\item{print}{If TRUE (default) plot is printed}
}
\value{
Plot of model performance as a function of algorithm and
  hyperparameter values tuned over. Generally called for the side effect of
  printing a plot, but the plot is also invisibly returned.
}
\description{
Plot performance of models
}
\examples{
plot(tune_models(mtcars, mpg))
}
