% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{plot.simulated_df}
\alias{plot.simulated_df}
\title{Plot Simulated Counterfactual Predictions}
\usage{
\method{plot}{simulated_df}(x, numeric_groups = 5,
  reorder_categories = TRUE, jitter_y = TRUE, x_var, color_var,
  font_size = 11, strip_size = 1, line_width = 0.5, print = TRUE, ...)
}
\arguments{
\item{x}{A simulated_df object from \code{link{simulate}}}

\item{numeric_groups}{For numeric variables that are converted to categories,
how many categories to create? Default = 5.}

\item{reorder_categories}{If TRUE (default) categorical variables that were
varied in \code{simulate} will be arranged in decreasing order of their
median predicted outcome}

\item{jitter_y}{If TRUE (default) and a variable is mapped to color (i.e. if
there is more than one varying variable), the vertical location of the
lines will be jittered slightly (no more than 1% of the range of the
outcome variable) to avoid overlap.}

\item{x_var}{Variable to put on the x-axis (unquoted). If not provided, the
most important variable is used, with numerics prioritized if one was
varied}

\item{color_var}{Variable to color lines (unquoted). If not provided, the
most important variable excluding \code{x_var}}

\item{...}{}
}
\details{

}
