% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-feature-availability-profiler.R
\name{plotProfiler}
\alias{plotProfiler}
\title{Display availability feature profile over time}
\usage{
plotProfiler(listOfVectors)
}
\arguments{
\item{listOfVectors}{A list of vectors, where first vector has the hours and
subsequent vectors represent the features and how much they're filled
for each of the hours. Usually populated by featureAvailabilityProfiler()}
}
\value{
Nothing
}
\description{
Shows what percentage of data is avilable after a particular 
starting time period.
}
\examples{
lis <- list()
# Establish hour range/sequence
lis$hoursSinceAdmit <- c(0,1,3,6,12,24)

# Add features and their percent full for each hour
lis$BP <- c(40, 45, 65, 78, 80, 90)
lis$LDL <- c(10, 30, 40, 70, 100, 120)

plotProfiler(lis)
}
\references{
\url{http://healthcareai-r.readthedocs.io}
}
\seealso{
\code{\link{healthcareai}}
}
