% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-local-interpretability.R
\name{plotVariableEffects}
\alias{plotVariableEffects}
\title{Plot the output of a model as each coefficient changes individually}
\usage{
plotVariableEffects(baseRow, 
                    modifiableCols, 
                    info,
                    info2,
                    predictFunction,
                    type,
                    spread = 2,
                    numberOfPercentiles = 12,
                    modifiableDfRow = NULL,
                    grid = NULL)
}
\arguments{
\item{baseRow}{a row in a data frame, used as the reference from which to 
vary the variables}

\item{modifiableCols}{a vector of column names corresponding to the columns
which should be perturbed}

\item{info}{a list which indexed by the column names in modifiableCols 
containing the standard deviation (for numeric columns) or the factor levels
(for factor columns)}

\item{info2}{list of percentiles TODO: expand}

\item{predictFunction}{a function with which to make new predictions for 
the model}

\item{type}{the type of fitObj: "classification" or "regression"}

\item{spread}{a number representing how widely to vary the variables}

\item{numberOfPercentiles}{the number of percentiles to use TODO:expand}

\item{modifiableDfRow}{the row in ModifiableFactorsDf corresponding to 
baseRow}

\item{grid}{An ordered pair of integers (rows, columns) which determines the 
number of plots to display at once. The first entry determines the number of
rows and the second determines the number of columns.}
}
\description{
Plot the output of a model as each coefficient changes 
individually
}
\references{
\url{http://healthcareai-r.readthedocs.io}
}
\seealso{
\code{\link{healthcareai}}
}
