% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.model_list}
\alias{predict.model_list}
\title{Make predictions}
\usage{
\method{predict}{model_list}(models, newdata, type = if (class(models[[1]]) ==
  "classification_list") "prob" else "raw", prepdata)
}
\arguments{
\item{models}{model_list object, as from `tune_models`}

\item{newdata}{data frame with same structure as the input to `tune_models`
that generated `models`. If the data that models were tuned on was
prepaired via `prep_data`, newdata will be prepared the same way, unless
prepdata is FALSE.}

\item{prepdata}{Logical, rarely needs to be set by the user. By default, if
`newdata` hasn't been prepped, it will be prepped by `prep_data` before
predictions are made. Set this to TRUE to force already-prepped data
through `prep_data` again, or set to FALSE to prevent `newdata` from being
sent through `prep_data`.}

\item{...}{Passed to `caret::predict.train`. For classification models,
\code{type = "raw"} will return Y/N predictions instead of class
probabilities.}
}
\value{
A tibble data frame: newdata with an additional column for the
  predictions in "predicted_TARGET" where TARGET is the name of the variable
  being predicted. The tibble will have child class "hcai_predicted_df" and
  attribute "model_info" that contains information about the model used to
  make predictions.
}
\description{
Make predictions
}
\details{
prepping data inside `predict` has the advantage of returning your
  predictions with the data frame in its original (unprepped) format. To do
  this, simply pass the data frame on which predictions are to be generated
  to `newdata` in the same format as the training data was passed to
  `prep_data` or `train_models`.
}
