% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{RemoveColsWithAllSameValue}
\alias{RemoveColsWithAllSameValue}
\title{Remove columns from a dataframe when those columns have the same values in each row}
\usage{
RemoveColsWithAllSameValue(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
A dataframe with those columns removed
}
\description{
Removes all of the columns in the dataframe that have the same value in every row
}
\examples{
df = data.frame(a=c(1,1,1),b=c('a','b','b'),c=c('a','a','a'),d = c(NA,'1',NA))
resdf = RemoveColsWithAllSameValue(df)
}
\seealso{
\code{\link{HCRTools}}
}

