% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{RemoveColsWithAllSameValue}
\alias{RemoveColsWithAllSameValue}
\title{Remove columns from a data frame when those columns have the same values in
each row}
\usage{
RemoveColsWithAllSameValue(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A data frame with zero-variance columns removed
}
\description{
Remove columns from a data frame when all of their rows are the
same value (after removing NA's)
}
\examples{
df = data.frame(a=c(1,1,1),
                b=c('a','b','b'),
                c=c('a','a','a'),
                d=c(NA,'1',NA))
df_result = RemoveColsWithAllSameValue(df)
head(df_result)
}
\references{
\url{http://healthcareml.org/}
}
\seealso{
\code{\link{HCRTools}}
}

