% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{RemoveRowsWithNAInSpecCol}
\alias{RemoveRowsWithNAInSpecCol}
\title{Remove rows where specified col is NA}
\usage{
RemoveRowsWithNAInSpecCol(df, desired_col)
}
\arguments{
\item{df}{A data frame to be altered}

\item{desired_col}{A column name in the df (in ticks)}
}
\value{
df_result The input data frame with rows removed
}
\description{
Remove rows from a data frame where a particular col is NA
}
\examples{
df = data.frame(a=c(1,2,3),b=c('Y','N',NA),c=c(NA,'Y','N'))
df_result = RemoveRowsWithNAInSpecCol(df,'b')
head(df_result)
}
\references{
\url{http://healthcareml.org/}
}
\seealso{
\code{\link{HCRTools}}
}

