% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-column-row-expectations.R
\name{removeRowsWithNAInSpecCol}
\alias{removeRowsWithNAInSpecCol}
\title{Remove rows where specified col is NA}
\usage{
removeRowsWithNAInSpecCol(df, desiredCol)
}
\arguments{
\item{df}{A data frame to be altered}

\item{desiredCol}{A column name in the df (in ticks)}
}
\value{
dfResult The input data frame with rows removed
}
\description{
Remove rows from a data frame where a particular col is NA
}
\examples{
df <- data.frame(a=c(1,2,3),b=c('Y','N',NA),c=c(NA,'Y','N'))
dfResult <- removeRowsWithNAInSpecCol(df,'b')
head(dfResult)
}
\references{
\url{http://healthcareai-r.readthedocs.io}
}
\seealso{
\code{\link{healthcareai}}
}
