% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiskAdjustedComparisons.R
\docType{class}
\name{RiskAdjustedComparisons}
\alias{RiskAdjustedComparisons}
\title{Make risk adjusted comparisons between groups/units or years/months}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RiskAdjustedComparisons
}
\arguments{
\item{df}{Dataframe whose columns are used for calc.}

\item{predicted.col}{Column that you want to predict.}

\item{group.col}{Column that we'll use to differentiate}

\item{impute}{Set all-column imputation to F or T.}
}
\description{
This class allows you to create a model based on the
performance of many groups in a cohort (besides group A, for example) and
see how well group A does against what the model would predict. Ranking each
of the groups this way provides a sense of which group's doing best in
terms of a particular measure.
}
\examples{

#### Example using csv data ####
library(HCRTools)
connection.string = "
driver={SQL Server};
server=localhost;
database=AdventureWorks2012;
trusted_connection=true
"

query = "
SELECT
[OrganizationLevel]
,[MaritalStatus]
,[Gender]
,IIF([SalariedFlag]=0,'N','Y') AS SalariedFlag
,[VacationHours]
,[SickLeaveHours]
FROM [AdventureWorks2012].[HumanResources].[Employee]
WHERE OrganizationLevel <> 0
"

df <- SelectData(connection.string, query)

p <- SupervisedModelParameters$new()
p$df = df
p$groupCol = 'OrganizationLevel'
p$impute = TRUE
p$predictedCol = 'SalariedFlag'
p$debug = FALSE
p$cores = 1

riskAdjComp <- RiskAdjustedComparisons$new(p)
riskAdjComp$run()

}
\references{
\url{http://healthcareml.org/}
}
\seealso{
\code{\link{HCRTools}}
}
\keyword{datasets}

