% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{SelectData}
\alias{SelectData}
\title{Pull data into R via an ODBC connection}
\usage{
SelectData(connection.string, query)
}
\arguments{
\item{connection.string}{A string specifying the driver, server, database,
and whether Windows Authentication will be used.}

\item{query}{The SQL query (in ticks)}
}
\value{
dataframe
}
\description{
Select data from an ODBC database and return the results as
a data frame.
}
\examples{
connection.string = '
  driver={SQL Server};
  server=localhost;
  database=AdventureWorks2012;
  trusted_connection=true'

query = '
  SELECT
    [OrganizationLevel]
  FROM [AdventureWorks2012].[HumanResources].[Employee]'

dataframe <- SelectData(connection.string, query)
}
\seealso{
\code{\link{HCRTools}}
}

