% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_train_test.R
\name{split_train_test}
\alias{split_train_test}
\title{Split data into training and test data frames}
\usage{
split_train_test(d, outcome, percent_train = 0.8, seed, grouping_col = NULL,
  aggreg_func = dplyr::first)
}
\arguments{
\item{d}{Data frame}

\item{outcome}{Target column, unquoted. Split will be stratified across this
variable}

\item{percent_train}{Proportion of rows in d to put into training. Default is 0.8}

\item{seed}{Optional, if provided the function will return the same split
each time it is called}

\item{grouping_col}{column name that specifies grouping. Individuals in the same
group are in the same training/test set.}

\item{aggreg_func}{when the data is grouped, the outcome column is aggregated
based on this function.}
}
\value{
A list of two data frames with names train and test
}
\description{
Split data into training and test data frames
}
\details{
This function wraps `caret::createDataPartition`. If outcome is a factor
  then the test/training porportions are stratified. Otherwise they are randomly
  selected.

  If the grouping_col is given, then the groups are divided into the test/
  training porportions by using the given aggregate function.
}
\examples{
split_train_test(mtcars, am, .9)
}
