% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingness.R
\name{summary.missingness}
\alias{summary.missingness}
\title{Summarizes data given by \code{\link{missingness}}}
\usage{
\method{summary}{missingness}(x, ...)
}
\arguments{
\item{x}{Data frame from \code{\link{missingness}}}
}
\value{
a tibble of the top 5 missingness percentage levels with the count of
  the number of variables associated with each level
}
\description{
Interpreting \code{\link{missingness}} results from wide
  datasets is difficult. This function helps interpret missingness output by
  summarizing this output by listing: 1. The percent of variables that
  contain missingness, 2. The variable name of the variable with the maximum
  amount of missingness along with its percent missingness, and 3. A tibble
  that lists the top 5 missingness levels with the count of the number of
  variables associated with each level.
}
\examples{
missingness(pima_diabetes) \%>\%
  summary()

}
