% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised-model.R
\docType{class}
\name{UnsupervisedModel}
\alias{UnsupervisedModel}
\title{Build clusters based on your data.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
UnsupervisedModel(object)
}
\arguments{
\item{object}{of UnsupervisedModelParams class for $new() constructor}
}
\description{
This step allows one to build clusters on your data
}
\references{
\url{http://healthcare.ai}
}
\seealso{
\code{\link{healthcareai}}

\code{\link{KmeansClustering}}
}
\keyword{datasets}
