% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-find-variation.R
\name{variationAcrossGroups}
\alias{variationAcrossGroups}
\title{Find variation across groups}
\usage{
variationAcrossGroups(df, categoricalCols, measureColumn, dateCol = NULL)
}
\arguments{
\item{df}{A dataframe}

\item{categoricalCols}{Vector of strings representing categorical column(s)}

\item{measureColumn}{Vector of strings representing measure column(s)}

\item{dateCol}{Optional. A date(time) column to group by (done by month)}
}
\value{
A boxplot to compare variation across groups.
}
\description{
Plot a boxplot to compare the variation across the groups.
}
\examples{
df <- data.frame(Dept = c('A','A','A','B','B','B','B','B'),
                 Gender = c('F','M','M','M','M','F','F','F'),
                 LOS = c(3.2,NA,5,1.3,2.4,4,9,10))

categoricalCols <- c("Dept","Gender")

variationAcrossGroups(df = df, 
                       categoricalCols = categoricalCols,
                       measureColumn = "LOS")

}
\references{
\url{http://healthcare.ai}
}
\seealso{
\code{\link{healthcareai}} \code{\link{calculateCOV}} 
\code{\link{createVarianceTallTable}}
}
