% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-find-variation.R
\name{variationAcrossGroups}
\alias{variationAcrossGroups}
\title{Find variation across groups}
\usage{
variationAcrossGroups(df, categoricalCols, measureColumn,
  printTukeyplot = FALSE, printTable = TRUE, boxplotStats = FALSE,
  dateCol = NULL)
}
\arguments{
\item{df}{A dataframe}

\item{categoricalCols}{Vector of strings representing categorical column(s)}

\item{measureColumn}{Vector of strings representing measure column(s)}

\item{printTukeyplot}{Optinal, default is FALSE. If TRUE, presents the plot 
returned by the Tukey's test.}

\item{printTable}{Optional, default is TRUE. FALSE: not to show the table of
mean/std and quartiles}

\item{boxplotStats}{Optinal, defalut is FALSE. If TRUE, returns the statistics
behindthe boxplot.}

\item{dateCol}{Optional. A date(time) column to group by (done by month)}
}
\value{
A boxplot to compare variation across groups.
}
\description{
Plot a boxplot to compare the variation across the groups.
}
\examples{

## Create the toy data set
df <- data.frame(Gender = factor(rbinom(100, 1, 0.45), labels = c("Male","Female")), 
                 Age = factor(rbinom(100, 1, 0.45), labels = c("Young","Old")),
                 Ans = factor(rbinom(100, 1, 0.45), labels = c("Y","N")),
                 Dept = sample(c("A","B","C"), 50, replace = TRUE, prob = c(0.2,0.3,0.5)),
                 measure1 = c(rnorm(30),rnorm(70,10,5)),
                 measure2 = c(rnorm(20,8,3),rnorm(80,-15,5)))

## Define the parameters 
categoricalCols <- c("Dept","Gender")
measureColumn <- c("measure1","measure2")

## Call the function
variationAcrossGroups(df = df, 
                       categoricalCols = categoricalCols,
                       measureColumn = measureColumn, printTukeyplot = TRUE)
                       
## Since printTukeyplot = TRUE and default of printTable is TRUE, the function 
## above will return the boxplot, the 95\% family-wise confidence interval plot, 
## as well as the tables. 
## In this example, the function returns
### Two boxplot
### 1. The boxplot of measure1 across the two factors, Dept and Gender
###    Dept has 3 levels: A, B, C
###    Gender has 2 levels: Male and Female
###    Hence, there are a total of 6 different levels if we consider both factors:
###    A.Male, B.Male, c.Male, A.Female, B.Female, C.Female
###    They are shown in the x axis of the boxplot. 
###    Levels that are not significantly different one each other are
###    represented with the same letter. The six boxes here all have letter "a",
###    meaning all the six levels here have the same mean.
### 2. The box plot of measure2 across the two factors, Dept and Gender
###    From this boxplot, we can conclude that the six levels have the 
###    same mean, or there is no significantly difference one each other
### 
### Two 95\% family-wise confidence level plots
###    There are the plots that present the results returned by Tukey's test. 
###    It allows to find means of a level that are significantly different 
###    from each other, comparing all possible pairs of means with a t-test 
###    like method. If for example, the means of the level of B.Frmale and B.Male
###    are significantly different, then the color of the line corresponding to
###    the pais is red, otherwise it's black. The order of the lines is according 
###    to the p-value, from the smallest to the largest. The notation in the 
###    topright of the plot is the measure column used to build the plot.
###    In this example, for both of the measure colums measure1 and measure, no
###    significant difference was found for any pair of the levels, which is 
###    the same as the boxplot.
### Two tables: one shows the p-values returned by the Tukey's test. And the 
### other presents the means/sd and quartiles of each level.

#########################################################################
set.seed(35)                      
treatment = c(rep("A", 20) , rep("B", 20) , rep("C", 20), rep("D", 20) ,  rep("E", 20))
value = c( sample(2:5, 20 , replace = TRUE) , sample(6:10, 20 , replace = TRUE), 
        sample(1:7, 20 , replace = TRUE), sample(3:10, 20 , replace = TRUE) , 
        sample(10:20, 20 , replace = TRUE) )
df1 = data.frame(treatment,value) ## This data set has two columns, one measure
                                  ## column value and one categorical column treatment.
                                  ## treatment has five levels: A B C D E
                                  
variationAcrossGroups(df1, categoricalCols = "treatment", measureColumn = "value")
## The boxplot tells us that treatment E has a significant different mean from 
## all the other treatments. Treatment A and C have no significant difference 
## in mean and so they share the same letter, and the same for treatment B and D. 
## But the means of treatment A and C are significant different from B,D.


}
\references{
\url{http://healthcare.ai}

\url{https://cran.r-project.org/web/packages/multcompView/index.html}
}
\seealso{
\code{\link{healthcareai}} \code{\link{findVariation}}
}
