% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-find-variation.R
\name{variationAcrossGroups}
\alias{variationAcrossGroups}
\title{Find variation across groups}
\usage{
variationAcrossGroups(df, categoricalCols, measureColumn,
  printTukeyplot = FALSE, printTable = TRUE, boxplotStats = FALSE,
  dateCol = NULL, levelOfDateGroup = "monthly")
}
\arguments{
\item{df}{A dataframe}

\item{categoricalCols}{Vector of strings representing categorical column(s)}

\item{measureColumn}{Vector of strings representing measure column(s)}

\item{printTukeyplot}{Optinal, default is FALSE. If TRUE, presents the plot 
returned by the Tukey's test.}

\item{printTable}{Optional, default is TRUE. FALSE: not to show the table of
mean/std and quartiles and the table of p values.}

\item{boxplotStats}{Optinal, defalut is FALSE. If TRUE, returns the statistics
behindthe boxplot.}

\item{dateCol}{Optional. A date(time) column to group by (done by month as default).}

\item{levelOfDateGroup}{Optional. Specify how to group the date(time) column,
can be yearly,quarterly,monthly(default) and weekly.}
}
\value{
A boxplot to compare variation across groups.
}
\description{
Plot a boxplot to compare the variation across the groups.
}
\examples{

## Create the toy data set
set.seed(2017)
n = 200
df <- data.frame(Dept = sample(c("Dept1","Dept2","Dept3"), 
                               size = n, 
                               replace = TRUE, 
                               prob =  c(0.5, 0.3, 0.2)),
                 Age = factor(rbinom(50, 1, 0.45), labels = c("Young","Old")),
                 Gender = factor(rbinom(50, 1, 0.45), labels = c("Male","Female")),
                 LOS = floor(abs(rnorm(n,10,5))),
                 BP = sample(50:180, n, replace = TRUE))

countA <- length(df$LOS[df$Dept == "Dept1"])
df$LOS[df$Dept == "Dept1"] <- floor(df$LOS[df$Dept == "Dept1"] + abs(rnorm(countA, mean = 2)))
countB <- length(df$LOS[df$Dept == "Dept2"])
df$LOS[df$Dept == "Dept2"] <- floor(df$LOS[df$Dept == "Dept2"] + abs(rnorm(countB, mean = 5)))
countY <- length(df$LOS[df$Age == "Young"])
df$LOS[df$Age == "Old"] <- floor(df$LOS[df$Age == "Old"] + abs(rnorm(countY, mean = 10)))
df$BP[df$Dept == "Dept1"] <- floor(df$BP[df$Dept == "Dept1"] + abs(rnorm(countA, mean = 2)))
df$BP[df$Dept == "Dept2"] <- floor(df$BP[df$Dept == "Dept2"] + abs(rnorm(countB, mean = 15)))
df$BP[df$Age == "Old"] <- floor(df$BP[df$Age == "Old"] + abs(rnorm(countY, mean = 20)))

head(df)

## Define the parameters 
categoricalCols <- c("Dept","Age")
measureColumn <- c("LOS","BP")

## Call the function
variationAcrossGroups(df = df, 
                       categoricalCols = categoricalCols,
                       measureColumn = measureColumn, printTukeyplot = TRUE)
                       
## Since printTukeyplot = TRUE and default of printTable is TRUE, the function 
## above will return the boxplot, the 95\% family-wise confidence interval plot, 
## as well as the tables. 
## In this example, the function returns
### Two boxplots
### 1. The boxplot of LOS across the two factors, Dept and Age
###    Dept has 3 levels: Dept1, Dept2, Dept3
###    Age has 2 levels: Young and Old
###    Hence, there are a total of 6 different levels if we consider both factors:
###    Dept1.Young, Dept2.Young, Dept3.Young, Dept1.Old, Dept2.Old, Dept3.Old
###    They are shown in the x axis of the boxplot. 
###    Levels that are not significantly different one each other are
###    represented with the same letter. For example, groups Dept1.Old and Dept3.Old do
###    not have a significant difference in mean.
### 2. The boxplot of BP across the two factors, Dept and Age
###    From this boxplot, we can conclude that the group Dept2.Old has a
###    significant higher mean.
### 
### Two 95\% family-wise confidence level plots
###    These are the plots that present the results returned by Tukey's test. 
###    It allows to find means of a level that are significantly different 
###    from each other, comparing all possible pairs of means with a t-test 
###    like method. Red lines indicate a significant difference between the pair
###    of groups. The order of the lines is according to the p-values, from the 
###    smallest to the largest. The notation in the bottomright of the plot is
###    the measure column used to build the plot.
### Two tables: one shows the p-values returned by the Tukey's test. And the 
### other presents the means, sd and quartiles of each group.

#########################################################################
set.seed(35)                      
treatment = c(rep("A", 50) , rep("B", 20) , rep("C", 40), rep("D", 80) ,  rep("E", 20))
value = c( sample(2:5, 50 , replace = TRUE) , sample(6:10, 20 , replace = TRUE), 
        sample(1:7, 40 , replace = TRUE), sample(3:10, 80 , replace = TRUE) , 
        sample(10:20, 20 , replace = TRUE) )
df1 = data.frame(treatment,value) ## This data set has two columns, one measure
                                  ## column value and one categorical column treatment.
                                  ## treatment has five levels: A B C D E
                                  
variationAcrossGroups(df1, categoricalCols = "treatment", measureColumn = "value")
## The boxplot tells us that treatment E has a significant different mean from 
## all the other treatments. Treatment A and C have no significant difference 
## in mean and so they share the same letter, and the same for treatment B and D. 
## But the means of treatment A and C are significantly different from B,D.


}
\references{
\url{http://healthcare.ai}

This function uses multcompLetters() in package multcompView
\url{https://cran.r-project.org/web/packages/multcompView/index.html}
}
\seealso{
\code{\link{healthcareai}} \code{\link{findVariation}}
}
