% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-find-variation.R
\name{variationAcrossGroups}
\alias{variationAcrossGroups}
\title{Find variation across groups}
\usage{
variationAcrossGroups(df, categoricalCols, measureColumn, printPlot = TRUE,
  printTable = TRUE, dateCol = NULL)
}
\arguments{
\item{df}{A dataframe}

\item{categoricalCols}{Vector of strings representing categorical column(s)}

\item{measureColumn}{Vector of strings representing measure column(s)}

\item{printPlot}{Optional, default is TRUE. FALSE: only shows the statistics
used to plot the boxplot.}

\item{printTable}{Optional, default is TRUE. FALSE: not to show the table of
mean/std and quartiles}

\item{dateCol}{Optional. A date(time) column to group by (done by month)}
}
\value{
A boxplot to compare variation across groups.
}
\description{
Plot a boxplot to compare the variation across the groups.
}
\examples{
df <- data.frame(Gender = factor(rbinom(100, 1, 0.45), labels = c("Male","Female")), 
                 Age = factor(rbinom(100, 1, 0.45), labels = c("Young","Old")),
                 Ans = factor(rbinom(100, 1, 0.45), labels = c("Y","N")),
                 Dept = sample(c("A","B","C"), 50, replace = TRUE, prob = c(0.2,0.3,0.5)),
                 LOS1 = c(rnorm(30),rnorm(70,10,5)),
                 LOS2 = c(rnorm(20,8,3),rnorm(80,-15,5)))

categoricalCols <- c("Dept","Gender")
measureColumn <- c("LOS1","LOS2")

variationAcrossGroups(df = df, 
                       categoricalCols = categoricalCols,
                       measureColumn = measureColumn)
                       
treatment = c(rep("A", 20) , rep("B", 20) , rep("C", 20), rep("D", 20) ,  rep("E", 20))
value = c( sample(2:5, 20 , replace = TRUE) , sample(6:10, 20 , replace = TRUE), 
        sample(1:7, 20 , replace = TRUE), sample(3:10, 20 , replace = TRUE) , 
        sample(10:20, 20 , replace = TRUE) )
df1 = data.frame(treatment,value)

variationAcrossGroups(df1, categoricalCols = "treatment", measureColumn = "value")

}
\references{
\url{http://healthcare.ai}
}
\seealso{
\code{\link{healthcareai}} \code{\link{calculateCOV}} 
\code{\link{createVarianceTallTable}}
}
