% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{WriteData}
\alias{WriteData}
\title{Write data to database}
\usage{
WriteData(df, server, database, schema_dot_table)
}
\arguments{
\item{df}{A data frame being written to a database}

\item{server}{A string.}

\item{database}{A string.}

\item{schema_dot_table}{A string representing the destination schema and
table. Note that brackets aren't expected.}
}
\value{
Nothing
}
\description{
Write data frame to database via ODBC connection
}
\examples{
library(HCRTools)
df <- data.frame(a=c(1,2,3),
                 b=c(2,4,6),
                 c=c('one','two','three'))

#WriteData(df,'localhost','SAM','dbo.HCRWriteData')
}
\references{
\url{https://community.healthcatalyst.com/community/data-science}
}
\seealso{
\code{\link{HCRTools}}
}

