% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{writeData}
\alias{writeData}
\title{Write data to database}
\usage{
writeData(df, server, database, schemaDotTable)
}
\arguments{
\item{df}{A data frame being written to a database}

\item{server}{A string.}

\item{database}{A string.}

\item{schemaDotTable}{A string representing the destination schema and
table. Note that brackets aren't expected.}
}
\value{
Nothing
}
\description{
Write data frame to database via ODBC connection
}
\examples{
df <- data.frame(a=c(1,2,3),
                 b=c(2,4,6),
                 c=c('one','two','three'))

writeData(df,'localhost','SAM','dbo.HCRWriteData')
}
\references{
\url{http://healthcare.ai}
}
\seealso{
\code{\link{HCRTools}}
}

