% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connections.R
\name{writeData}
\alias{writeData}
\title{Defunct. See
\href{https://docs.healthcare.ai/articles/site_only/db_connections.html}{this
vignette} for help writing to databases.}
\usage{
writeData(...)
}
\arguments{
\item{...}{Garbage collector}
}
\description{
Removed in v2.0.0
}
